Refatoração
Separado sistema em Main, ST, e UI
