+++ utilitarios/yo/element/tree/Actions.java
@@ -20,7 +20,6 @@ import java.io.StringReader;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import javax.swing.JFrame;
 import javax.xml.parsers.DocumentBuilderFactory;
@@ -31,120 +30,14 @@ import element.tree.objeto.ListaObjeto;
 import element.tree.objeto.Objeto;
 import element.tree.objeto.conexao.Conexao;
 import element.tree.propriedades.Icone;
-import element.tree.objeto.conexao.Nodulo;
-import element.tree.objeto.conexao.Segmento;
+import element.tree.objeto.nodulo.Nodulo;
+import element.tree.objeto.conexao.segmento.Segmento;
 import element.tree.objeto.modulo.Modulo;
 import element.tree.undoRedo.UndoRedoObjeto;
 public class Actions{
-//STATES
-	public final static int NORMAL=0;				//[0-9]
-	public final static int TO_CREATE=1;			//[0-9]
-	public final static int TO_CONNECT=2;			//[0-9]
-	public final static int TO_DELETE=3;			//[0-9]
-	public final static int EDIT_TITLE=4;			//[0-9]
-	public final static int DRAG=10;				//[0-9]0
-	public final static int MOVE=20;				//[0-9]0
-	public final static int SELECT_AREA=30;			//[0-9]0
-	public final static int WAITING1=100;			//[0-9]00
-	public final static int WAITING2=200;			//[0-9]00
-	public final static int WAITING3=300;			//[0-9]00
-	public final static int LEFT=1000;				//[0-9]000
-	public final static int MIDDLE=2000;			//[0-9]000
-	public final static int RIGHT=3000;				//[0-9]000
-	public final static int DRAG_ALL=10000;			//[0-9]0000
-	public final static int AUTO_DRAG_ALL=20000;	//[0-9]0000
-	private int state=NORMAL;
-		public int getState(){return state;}
-	private class State{
-		private Action mousePressed;
-			public Action getMousePressedAction(){return mousePressed;}
-			public void setMousePressedAction(Action action){mousePressed=action;}
-		private Action mouseReleased;
-			public Action getMouseReleasedAction(){return mouseReleased;}
-			public void setMouseReleasedAction(Action action){mouseReleased=action;}
-		private Action mouseCliqued;
-			public Action getMouseCliquedAction(){return mouseCliqued;}
-			public void setMouseCliquedAction(Action action){mouseCliqued=action;}
-		private Action mouseDragged;
-			public Action getMouseDraggedAction(){return mouseDragged;}
-			public void setMouseDraggedAction(Action action){mouseDragged=action;}
-		private Action mouseMoved;
-			public Action getMouseMovedAction(){return mouseMoved;}
-			public void setMouseMovedAction(Action action){mouseMoved=action;}
-		private Action keyPressed;
-			public Action getKeyPressedAction(){return keyPressed;}
-			public void setKeyPressedAction(Action action){keyPressed=action;}
-		private Action keyReleased;
-			public Action getKeyReleasedAction(){return keyReleased;}
-			public void setKeyReleasedAction(Action action){keyReleased=action;}
-		private Action created;
-			public Action getCreateAction(){return created;}
-			public void setCreateAction(Action action){created=action;}
-		private Action connected;
-			public Action getConnectAction(){return connected;}
-			public void setConnectAction(Action action){connected=action;}
-		private Action deleted;
-			public Action getDeleteAction(){return deleted;}
-			public void setDeleteAction(Action action){deleted=action;}
-		private Modulo mod;
-			public void setModulo(Modulo mod){this.mod=mod;}
-			public Modulo getModulo(){return mod;}
-		private Conexao cox;
-			public void setConexao(Conexao cox){this.cox=cox;}
-			public Conexao getConexao(){return cox;}
-		private Nodulo nod;
-			public void setNodulo(Nodulo nod){this.nod=nod;}
-			public Nodulo getNodulo(){return nod;}
-		private Segmento seg;
-			public void setSegmento(Segmento seg){this.seg=seg;}
-			public Segmento getSegmento(){return seg;}
-		private Action action;
-			public void setAction(Action action){this.action=action;}
-			public Action getAction(){return action;}
-	}
-	private abstract class Action{
-		public void run(){}
-		public void run(MouseEvent m){}
-		public void run(KeyEvent k){}
-	}
-	private HashMap<Integer,State>states=new HashMap<>();
-		private HashMap<Integer,State>getStates(){return states;}
-		public State setState(int newStateIndex,Objeto...objs){
-			final State oldState=getStates().get(state);
-			oldState.setModulo(null);		//ZERA MOD
-			oldState.setConexao(null);		//ZERA COX
-			oldState.setNodulo(null);		//ZERA NOD
-			oldState.setSegmento(null);		//ZERA SEG
-			state=newStateIndex;
-			final State newState=getStates().get(newStateIndex);
-			if(objs.length>0)for(Objeto obj:objs){
-				switch(obj.getTipo()){
-					case MODULO:	newState.setModulo((Modulo)obj);		break;
-					case CONEXAO:	newState.setConexao((Conexao)obj);		break;
-					case NODULO:	newState.setNodulo((Nodulo)obj);		break;
-					case SEGMENTO:	newState.setSegmento((Segmento)obj);	break;
-				}
-			}
-			if(newState.getAction()!=null)newState.getAction().run();
-			return newState;
-		}
-		public boolean stateContains(int state){
-			final int camada1=(state%10000%1000%100%10);
-			if(state==camada1)return true;
-			final int camada2=(state%10000%1000%100)-camada1;
-			if(state==camada2)return true;
-			final int camada3=(state%10000%1000)-camada1-camada2;
-			if(state==camada3)return true;
-			final int camada4=(state%10000)-camada1-camada2-camada3;
-			if(state==camada4)return true;
-			final int camada5=(state)-camada1-camada2-camada3-camada4;
-			if(state==camada5)return true;
-			return false;
-		}
 //VAR GLOBAIS
-	private JFrame janela;
 	private Tree tree;
-	private Point oldMouse=new Point();
+	protected Point oldMouse=new Point();
 //MOUSES
 	public Point mousePressed=new Point();
 	public Point mouseReleased=new Point();
@@ -152,2904 +45,12 @@ public class Actions{
 	public Point mouseMoved=new Point();
 	public Point mouseNonGrid=new Point();
 //MAIN
-	public Actions(Tree tree,JFrame janela){
+	public Actions(Tree tree){
 		this.tree=tree;
-		this.janela=janela;
-//NORMAL
-		states.put(NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.NORMAL);
-				};
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Modulo modSelec=getModuloHover(mouse);
-						if(modSelec!=null){										//HÁ MOD SELEC
-							if(modSelec.getState().is(Modulo.State.SELECTED)){
-								setState(LEFT+WAITING2+NORMAL,modSelec);
-							}else{
-								setState(LEFT+WAITING1+NORMAL,modSelec);
-							}
-						}else{													//NENHUM MOD FOI SELEC, VERIFICA POR NODS
-							final Nodulo nodSelec=getNoduloHover(mouse);
-							if(nodSelec!=null){									//HÁ NOD SELEC
-								if(nodSelec.getState().is(Nodulo.State.SELECTED)){
-									setState(LEFT+WAITING2+NORMAL,nodSelec);
-								}else{
-									setState(LEFT+WAITING1+NORMAL,nodSelec);
-								}
-							}else{												//NENHUM NOD FOI SELEC, VERIFICA POR COXS
-								final Conexao coxSelec=getConexaoHover(mouse);
-								if(coxSelec!=null){								//HÁ COX SELEC
-									if(coxSelec.getState().is(Conexao.State.SELECTED)){
-										setState(LEFT+WAITING2+NORMAL,coxSelec);
-									}else{
-										setState(LEFT+WAITING1+NORMAL,coxSelec);
-									}
-								}else{											//NENHUM COX FOI SELEC
-									setState(LEFT+WAITING1+NORMAL);
-								}
-							}
-						}
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(MIDDLE+WAITING1+NORMAL);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseCliquedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)&&m.getClickCount()==2){
-						final Point mouse=getPosition(m.getPoint());
-						final Modulo modSelec=getModuloHover(mouse);
-						if(modSelec!=null){
-							tree.unSelectAll();
-							tree.select(modSelec);
-							tree.draw();
-							tree.setTitulo(modSelec);	//TITULO ACIMA DO DRAW
-							setState(EDIT_TITLE);
-						}
-					}
-				}
-			});
-			setKeyPressedAction(new Action(){
-				public void run(KeyEvent k){
-					if(tree.getPopup().isShowing())return;
-					boolean ctrl=(k.isControlDown());
-					final int salto=(ctrl?8:1);
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_UP:
-						case KeyEvent.VK_KP_UP:
-							if(tree.getSelectedObjetos().isEmpty()){
-								Tree.setLocal(Tree.getLocalX(),Tree.getLocalY()+salto);
-							}else{
-								for(Modulo mod:tree.getSelectedObjetos().getModulos()){
-									mod.setLocationIndex(mod.getXIndex(),mod.getYIndex()-salto);
-								}
-								for(Nodulo nod:tree.getSelectedObjetos().getNodulos()){
-									nod.setLocationIndex(nod.getXIndex(),nod.getYIndex()-salto);
-								}
-								final List<Objeto>savedObjs=new ArrayList<>();
-								savedObjs.addAll(tree.getSelectedObjetos().getModulos());
-								savedObjs.addAll(tree.getSelectedObjetos().getNodulos());
-								tree.getUndoRedoManager().addUndoLocal(savedObjs.toArray(new Objeto[0]),new Point(0,+salto));
-							}
-							tree.draw();
-						break;
-						case KeyEvent.VK_DOWN:
-						case KeyEvent.VK_KP_DOWN:
-							if(tree.getSelectedObjetos().isEmpty()){
-								Tree.setLocal(Tree.getLocalX(),Tree.getLocalY()-salto);
-							}else{
-								for(Modulo mod:tree.getSelectedObjetos().getModulos()){
-									mod.setLocationIndex(mod.getXIndex(),mod.getYIndex()+salto);
-								}
-								for(Nodulo nod:tree.getSelectedObjetos().getNodulos()){
-									nod.setLocationIndex(nod.getXIndex(),nod.getYIndex()+salto);
-								}
-								final List<Objeto>savedObjs=new ArrayList<>();
-								savedObjs.addAll(tree.getSelectedObjetos().getModulos());
-								savedObjs.addAll(tree.getSelectedObjetos().getNodulos());
-								tree.getUndoRedoManager().addUndoLocal(savedObjs.toArray(new Objeto[0]),new Point(0,-salto));
-							}
-							tree.draw();
-						break;
-						case KeyEvent.VK_RIGHT:
-						case KeyEvent.VK_KP_RIGHT:
-							if(tree.getSelectedObjetos().isEmpty()){
-								Tree.setLocal(Tree.getLocalX()-salto,Tree.getLocalY());
-							}else{
-								for(Modulo mod:tree.getSelectedObjetos().getModulos()){
-									mod.setLocationIndex(mod.getXIndex()+salto,mod.getYIndex());
-								}
-								for(Nodulo nod:tree.getSelectedObjetos().getNodulos()){
-									nod.setLocationIndex(nod.getXIndex()+salto,nod.getYIndex());
-								}
-								final List<Objeto>savedObjs=new ArrayList<>();
-								savedObjs.addAll(tree.getSelectedObjetos().getModulos());
-								savedObjs.addAll(tree.getSelectedObjetos().getNodulos());
-								tree.getUndoRedoManager().addUndoLocal(savedObjs.toArray(new Objeto[0]),new Point(-salto,0));
-							}
-							tree.draw();
-						break;
-						case KeyEvent.VK_LEFT:
-						case KeyEvent.VK_KP_LEFT:
-							if(tree.getSelectedObjetos().isEmpty()){
-								Tree.setLocal(Tree.getLocalX()+salto,Tree.getLocalY());
-							}else{
-								for(Modulo mod:tree.getSelectedObjetos().getModulos()){
-									mod.setLocationIndex(mod.getXIndex()-salto,mod.getYIndex());
-								}
-								for(Nodulo nod:tree.getSelectedObjetos().getNodulos()){
-									nod.setLocationIndex(nod.getXIndex()-salto,nod.getYIndex());
-								}
-								final List<Objeto>savedObjs=new ArrayList<>();
-								savedObjs.addAll(tree.getSelectedObjetos().getModulos());
-								savedObjs.addAll(tree.getSelectedObjetos().getNodulos());
-								tree.getUndoRedoManager().addUndoLocal(savedObjs.toArray(new Objeto[0]),new Point(salto,0));
-							}
-							tree.draw();
-						break;
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE);
-					setState(TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE_SON);
-					setState(TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_DELETE);
-					setState(TO_DELETE);
-				}
-			});
-		}});
-//MIDDLE+WAITING1+NORMAL
-		states.put(MIDDLE+WAITING1+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.SELECT);
-				};
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(AUTO_DRAG_ALL+NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-						setState(DRAG_ALL+NORMAL);
-					}
-				}
-			});
-		}});
-//DRAG_ALL+NORMAL
-		states.put(DRAG_ALL+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.SELECT);
-				};
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE);
-					setState(DRAG_ALL+TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE_SON);
-					setState(DRAG_ALL+TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_DELETE);
-					setState(DRAG_ALL+TO_DELETE);
-				}
-			});
-		}});
-//AUTO_DRAG_ALL+NORMAL
-		states.put(AUTO_DRAG_ALL+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.AUTODRAG);
-					startAutoDrag();
-				};
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						setState(NORMAL);
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(MIDDLE+WAITING1+NORMAL);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						setState(NORMAL);
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE);
-					setState(TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE_SON);
-					setState(TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_DELETE);
-					setState(TO_DELETE);
-				}
-			});
-		}});
-//LEFT+WAITING1+NORMAL
-		states.put(LEFT+WAITING1+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.SELECT);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+LEFT+WAITING1+NORMAL,getModulo(),getConexao(),getNodulo());
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						if(getModulo()!=null){				//HÁ MOD SELEC
-							if(getModulo().contains(mouse))tree.select(getModulo());
-						}else if(getNodulo()!=null){		//HÁ NOD SELEC
-							if(getNodulo().contains(mouse))tree.select(getNodulo());
-						}else if(getConexao()!=null){		//HÁ COX SELEC
-							if(getConexao().contains(mouse))tree.select(getConexao());
-						}
-						tree.draw();
-						setState(NORMAL);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						tree.unSelectAll();
-						final Point mouse=getPosition(m.getPoint());
-						if(getModulo()!=null){				//HÁ MOD SELEC
-							if(getModulo().contains(mouse))tree.select(getModulo());
-						}else if(getNodulo()!=null){		//HÁ NOD SELEC
-							if(getNodulo().contains(mouse))tree.select(getNodulo());
-						}else if(getConexao()!=null){		//HÁ COX SELEC
-							if(getConexao().contains(mouse))tree.select(getConexao());
-						}
-						tree.draw();
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.ALT,Cursor.LEFT)){
-						if(getConexao()!=null){
-							for(Conexao cox:tree.getSelectedObjetos().getConexoes()){
-								for(Nodulo nod:cox.getNodulos()){
-									tree.select(nod);
-								}
-							}
-						}
-						oldMouse=new Point(mouseDragged);
-						for(Modulo mod:tree.getSelectedObjetos().getModulos()){
-							mod.setLocationIndex(mod.getXIndex()-(mouseDragged.x-mouseDraggedAtual.x),mod.getYIndex()-(mouseDragged.y-mouseDraggedAtual.y));
-						}
-						for(Nodulo nod:tree.getSelectedObjetos().getNodulos()){
-							nod.setLocationIndex(nod.getXIndex()-(mouseDragged.x-mouseDraggedAtual.x),nod.getYIndex()-(mouseDragged.y-mouseDraggedAtual.y));
-						}
-						tree.draw();
-						setState(DRAG+NORMAL);
-					
-					}else if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						setArea(Selecao.State.TO_SELECT,true,
-								mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-								mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						setState(SELECT_AREA+NORMAL);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						setArea(Selecao.State.TO_SELECT,false,
-								mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-								mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						setState(SELECT_AREA+NORMAL);
-					}
-				}
-			});
-		}});
-//DRAG_ALL+LEFT+WAITING1+NORMAL
-		states.put(DRAG_ALL+LEFT+WAITING1+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						if(getModulo()!=null){				//HÁ MOD SELEC
-							if(getModulo().contains(mouse))tree.select(getModulo());
-						}else{
-							if(getNodulo()!=null){			//HÁ NOD SELEC
-								if(getNodulo().contains(mouse))tree.select(getNodulo());
-							}else{
-								if(getConexao()!=null){		//HÁ COX SELEC
-									if(getConexao().contains(mouse))tree.select(getConexao());
-								}
-							}
-						}
-						tree.draw();
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						tree.unSelectAll();
-						final Point mouse=getPosition(m.getPoint());
-						if(getModulo()!=null){				//HÁ MOD SELEC
-							if(getModulo().contains(mouse))tree.select(getModulo());
-						}else{
-							if(getNodulo()!=null){			//HÁ NOD SELEC
-								if(getNodulo().contains(mouse))tree.select(getNodulo());
-							}else{
-								if(getConexao()!=null){		//HÁ COX SELEC
-									if(getConexao().contains(mouse))tree.select(getConexao());
-								}
-							}
-						}
-						tree.draw();
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						setState(LEFT+WAITING1+NORMAL,getModulo(),getConexao(),getNodulo());
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						updateTituloBounds();
-						tree.draw();
-					}
-				}
-			});
-		}});
-//SELECT_AREA+NORMAL
-		states.put(SELECT_AREA+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.AREA_SELECT+Cursor.SELECT);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						setState(DRAG_ALL+SELECT_AREA+NORMAL);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						tree.getSelecao().setEmpty();
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						tree.getSelecao().setEmpty();
-						tree.draw();
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						setArea(Selecao.State.TO_SELECT,true,
-								mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						setArea(Selecao.State.TO_SELECT,false,
-								mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-					}
-				}
-			});
-			setKeyReleasedAction(new Action(){
-				public void run(KeyEvent k){
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_CONTROL:
-							setArea(Selecao.State.TO_SELECT,false);
-						break;
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setArea(Selecao.State.TO_CREATE,false);
-					setState(SELECT_AREA+TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setArea(Selecao.State.TO_CREATE_SON,false);
-					setState(SELECT_AREA+TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setArea(Selecao.State.TO_DELETE,false);
-					setState(SELECT_AREA+TO_DELETE);
-				}
-			});
-		}});
-//DRAG_ALL+SELECT_AREA+NORMAL
-		states.put(DRAG_ALL+SELECT_AREA+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.AREA_SELECT+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						tree.getSelecao().setEmpty();
-						tree.draw();
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(SELECT_AREA+NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-			setKeyReleasedAction(new Action(){
-				public void run(KeyEvent k){
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_CONTROL:
-							setArea(Selecao.State.TO_SELECT,false);
-						break;
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setArea(Selecao.State.TO_CREATE,false);
-					setState(DRAG_ALL+SELECT_AREA+TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setArea(Selecao.State.TO_CREATE_SON,false);
-					setState(DRAG_ALL+SELECT_AREA+TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setArea(Selecao.State.TO_DELETE,false);
-					setState(DRAG_ALL+SELECT_AREA+TO_DELETE);
-				}
-			});
-		}});
-//LEFT+WAITING2+NORMAL
-		states.put(LEFT+WAITING2+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.SELECT);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						setState(DRAG_ALL+LEFT+WAITING2+NORMAL,getModulo(),getConexao(),getNodulo());
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						if(getModulo()!=null){				//HÁ MOD SELEC
-							if(getModulo().contains(mouse))tree.unSelect(getModulo());
-						}else if(getNodulo()!=null){		//HÁ NOD SELEC
-							if(getNodulo().contains(mouse))tree.unSelect(getNodulo());
-						}else if(getConexao()!=null){		//HÁ COX SELEC
-							if(getConexao().contains(mouse))tree.unSelect(getConexao());
-						}
-						tree.draw();
-						setState(NORMAL);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						if(tree.getSelectedObjetos().getAll().size()==1){
-							tree.unSelectAll();
-						}else{
-							tree.unSelectAll();
-							if(getModulo()!=null){
-								if(getModulo().contains(mouse))tree.select(getModulo());
-							}else if(getNodulo()!=null){
-								if(getNodulo().contains(mouse))tree.select(getNodulo());
-							}else if(getConexao()!=null){
-								if(getConexao().contains(mouse))tree.select(getConexao());
-							}
-						}
-						tree.draw();
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						setArea(Selecao.State.TO_SELECT,true,
-								mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-								mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						setState(SELECT_AREA+NORMAL);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						if(getConexao()!=null){
-							for(Conexao cox:tree.getSelectedObjetos().getConexoes()){
-								for(Nodulo nod:cox.getNodulos()){
-									tree.select(nod);
-								}
-							}
-						}
-						oldMouse=new Point(mouseDragged);
-						for(Modulo mod:tree.getSelectedObjetos().getModulos()){
-							mod.setLocationIndex(mod.getXIndex()-(mouseDragged.x-mouseDraggedAtual.x),mod.getYIndex()-(mouseDragged.y-mouseDraggedAtual.y));
-						}
-						for(Nodulo nod:tree.getSelectedObjetos().getNodulos()){
-							nod.setLocationIndex(nod.getXIndex()-(mouseDragged.x-mouseDraggedAtual.x),nod.getYIndex()-(mouseDragged.y-mouseDraggedAtual.y));
-						}
-						tree.draw();
-						setState(DRAG+NORMAL);
-					}
-				}
-			});
-		}});
-//DRAG_ALL+LEFT+WAITING2+NORMAL
-		states.put(DRAG_ALL+LEFT+WAITING2+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						if(getModulo()!=null){				//HÁ MOD SELEC
-							if(getModulo().contains(mouse))tree.unSelect(getModulo());
-						}else{
-							if(getNodulo()!=null){			//HÁ NOD SELEC
-								if(getNodulo().contains(mouse))tree.unSelect(getNodulo());
-							}else{
-								if(getConexao()!=null){		//HÁ COX SELEC
-									if(getConexao().contains(mouse))tree.unSelect(getConexao());
-								}
-							}
-						}
-						tree.draw();
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						if(getModulo()!=null){
-							if(getModulo().contains(mouse)){
-								if(tree.getSelectedObjetos().getModulos().size()>1){
-									 tree.unSelectAll();
-									 tree.select(getModulo());
-								}else tree.unSelectAll();
-							}else tree.unSelectAll();
-						}else{
-							if(getNodulo()!=null){
-								if(getNodulo().contains(mouse)){
-									if(tree.getSelectedObjetos().getNodulos().size()>1){
-										 tree.unSelectAll();
-										 tree.select(getNodulo());
-									}else tree.unSelectAll();
-								}else tree.unSelectAll();
-							}else{
-								if(getConexao()!=null){
-									if(getConexao().contains(mouse)){
-										if(tree.getSelectedObjetos().getConexoes().size()>1){
-											 tree.unSelectAll();
-											 tree.select(getConexao());
-										}else tree.unSelectAll();
-									}else tree.unSelectAll();
-								}else tree.unSelectAll();	//SE CLICK NO VAZIO
-							}
-						}
-						tree.draw();
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(LEFT+WAITING2+NORMAL,getModulo(),getConexao(),getNodulo());
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						updateTituloBounds();
-						tree.draw();
-					}
-				}
-			});
-		}});
-//DRAG+NORMAL
-		states.put(DRAG+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.MOVE+Cursor.SELECT);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+DRAG+NORMAL);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						final List<Objeto>savedObjs=new ArrayList<>();
-						savedObjs.addAll(tree.getSelectedObjetos().getModulos());
-						savedObjs.addAll(tree.getSelectedObjetos().getNodulos());
-						tree.getUndoRedoManager().addUndoLocal(savedObjs.toArray(new Objeto[0]),new Point(-(mouseMoved.x-oldMouse.x),-(mouseMoved.y-oldMouse.y)));
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						final List<Objeto>savedObjs=new ArrayList<>();
-						savedObjs.addAll(tree.getSelectedObjetos().getModulos());
-						savedObjs.addAll(tree.getSelectedObjetos().getNodulos());
-						tree.getUndoRedoManager().addUndoLocal(savedObjs.toArray(new Objeto[0]),new Point(-(mouseMoved.x-oldMouse.x),-(mouseMoved.y-oldMouse.y)));
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT)){
-						for(Modulo mod:tree.getSelectedObjetos().getModulos()){
-							mod.setLocationIndex(mod.getXIndex()-(mouseDragged.x-mouseDraggedAtual.x),mod.getYIndex()-(mouseDragged.y-mouseDraggedAtual.y));
-						}
-						for(Nodulo nod:tree.getSelectedObjetos().getNodulos()){
-							nod.setLocationIndex(nod.getXIndex()-(mouseDragged.x-mouseDraggedAtual.x),nod.getYIndex()-(mouseDragged.y-mouseDraggedAtual.y));
-						}
-						tree.draw();
-					}
-				}
-			});
-		}});
-//DRAG_ALL+DRAG+NORMAL
-		states.put(DRAG_ALL+DRAG+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.MOVE+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						final List<Objeto>savedObjs=new ArrayList<>();
-						savedObjs.addAll(tree.getSelectedObjetos().getModulos());
-						savedObjs.addAll(tree.getSelectedObjetos().getNodulos());
-						tree.getUndoRedoManager().addUndoLocal(savedObjs.toArray(new Objeto[0]),new Point(-(mouseMoved.x-oldMouse.x),-(mouseMoved.y-oldMouse.y)));
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG+NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-						if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-		}});
-//MOVE+NORMAL
-		states.put(MOVE+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.MOVE+Cursor.NORMAL);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						final List<Objeto>savedObjs=new ArrayList<>(tree.getSelectedObjetos().getAll().values());
-						addUndoable(savedObjs,false);
-						setState(NORMAL);
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+MOVE+NORMAL);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						final List<Objeto>savedObjs=new ArrayList<>(tree.getSelectedObjetos().getAll().values());
-						addUndoable(savedObjs,false);
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseMovedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseMovedAtual=getGridPosition(m.getPoint());
-					for(Modulo mod:tree.getSelectedObjetos().getModulos()){
-						mod.setLocationIndex(mod.getXIndex()-(mouseMoved.x-mouseMovedAtual.x),mod.getYIndex()-(mouseMoved.y-mouseMovedAtual.y));
-					}
-					for(Nodulo nod:tree.getSelectedObjetos().getNodulos()){
-						nod.setLocationIndex(nod.getXIndex()-(mouseMoved.x-mouseMovedAtual.x),nod.getYIndex()-(mouseMoved.y-mouseMovedAtual.y));
-					}
-					tree.draw();
-				}
-			});
-		}});
-//DRAG_ALL+MOVE+NORMAL
-		states.put(DRAG_ALL+MOVE+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.MOVE+Cursor.SELECT);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						final List<Objeto>savedObjs=new ArrayList<>(tree.getSelectedObjetos().getAll().values());
-						addUndoable(savedObjs,false);
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						final List<Objeto>savedObjs=new ArrayList<>(tree.getSelectedObjetos().getAll().values());
-						addUndoable(savedObjs,false);
-						tree.unSelectAll();
-						tree.draw();
-						setState(DRAG_ALL+RIGHT+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(MOVE+NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-						if(Cursor.match(m,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-		}});
-//EDIT_TITLE
-	states.put(EDIT_TITLE,new State(){{
-		setAction(new Action(){
-			public void run(){
-				tree.getCursor().set(Cursor.EDIT_TITLE+Cursor.NORMAL);
-			}
-		});
-		setMousePressedAction(new Action(){
-			public void run(MouseEvent m){
-				if(Cursor.match(m,Cursor.LEFT)){
-					tree.getCursor().set(Cursor.EDIT_TITLE+Cursor.SELECT);
-				}else if(Cursor.match(m,Cursor.MIDDLE)){
-					setState(DRAG_ALL+EDIT_TITLE);
-				}else if(Cursor.match(m,Cursor.RIGHT)){
-					tree.setTitulo(null);
-					tree.draw();
-					setState(RIGHT+WAITING1+NORMAL);
-				}
-			}
-		});
-		setMouseReleasedAction(new Action(){
-			public void run(MouseEvent m){
-				if(Cursor.match(m,Cursor.LEFT)){
-					tree.setTitulo(null);
-					tree.draw();
-					setState(NORMAL);
-				}
-			}
-		});
-		setMouseDraggedAction(new Action(){
-			public void run(MouseEvent m){
-				final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-				if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-					tree.setTitulo(null);
-					setArea(Selecao.State.TO_SELECT,true,
-							mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-							mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-					setState(SELECT_AREA+NORMAL);
-				}else if(Cursor.match(m,Cursor.LEFT)){
-					tree.setTitulo(null);
-					setArea(Selecao.State.TO_SELECT,false,
-							mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-							mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-					setState(SELECT_AREA+NORMAL);
-				}
-			}
-		});
-		setKeyReleasedAction(new Action(){
-			public void run(KeyEvent k){
-				switch(k.getKeyCode()){
-					case KeyEvent.VK_ESCAPE:
-						//delete() É ATIVADO
-						setState(NORMAL);
-					break;
-				}
-			}
-		});
-	}});
-//DRAG_ALL+EDIT_TITLE
-	states.put(DRAG_ALL+EDIT_TITLE,new State(){{
-		setAction(new Action(){
-			public void run(){
-				tree.getCursor().set(Cursor.DRAG+Cursor.EDIT_TITLE+Cursor.SELECT);
-			}
-		});
-		setMousePressedAction(new Action(){
-			public void run(MouseEvent m){
-				if(Cursor.match(m,Cursor.MIDDLE)){
-					setState(DRAG_ALL+EDIT_TITLE);
-				}else if(Cursor.match(m,Cursor.RIGHT)){
-					tree.setTitulo(null);
-					tree.draw();
-					setState(RIGHT+WAITING1+NORMAL);
-				}
-			}
-		});
-		setMouseReleasedAction(new Action(){
-			public void run(MouseEvent m){
-				if(Cursor.match(m,Cursor.LEFT)){
-					tree.setTitulo(null);
-					tree.draw();
-					setState(DRAG_ALL+NORMAL);
-				}else if(Cursor.match(m,Cursor.MIDDLE)){
-					setState(EDIT_TITLE);
-				}
-			}
-		});
-		setMouseDraggedAction(new Action(){
-			public void run(MouseEvent m){
-				final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-				if(Cursor.match(m,Cursor.MIDDLE)){
-					Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-					updateTituloBounds();
-					tree.draw();
-				}
-			}
-		});
-		setKeyReleasedAction(new Action(){
-			public void run(KeyEvent k){
-				switch(k.getKeyCode()){
-					case KeyEvent.VK_ESCAPE:
-						tree.setTitulo(null);
-						tree.draw();
-						setState(DRAG_ALL+NORMAL);
-					break;
-				}
-			}
-		});
-	}});
-//RIGHT+WAITING1+NORMAL
-		states.put(RIGHT+WAITING1+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.RIGHT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Modulo modSelec=getModuloHover(mouse);
-						if(modSelec!=null){
-							if(!modSelec.getState().is(Modulo.State.SELECTED))tree.unSelectAll();
-							tree.unSelect(modSelec);
-							tree.select(modSelec);
-							tree.getPopup().show(mousePressed,modSelec);
-						}else{
-							final Nodulo nodSelec=getNoduloHover(mouse);
-							if(nodSelec!=null){
-								if(!nodSelec.getState().is(Nodulo.State.SELECTED))tree.unSelectAll();
-								tree.unSelect(nodSelec);
-								tree.select(nodSelec);
-								tree.getPopup().show(mousePressed,nodSelec);
-							}else{
-								final Conexao coxSelec=getConexaoHover(mouse);
-								if(coxSelec!=null){
-									if(!coxSelec.getState().is(Conexao.State.SELECTED))tree.unSelectAll();
-									tree.unSelect(coxSelec);
-									tree.select(coxSelec);
-									final Segmento segSelec=coxSelec.segmentContains(mouse);
-									tree.getPopup().show(mousePressed,segSelec);
-								}else{
-									tree.unSelectAll();
-									tree.getPopup().show(mousePressed,null);
-								}
-							}
-						}
-						tree.draw();
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.RIGHT)){
-						tree.unSelectAll();
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-						setState(DRAG_ALL+RIGHT+NORMAL);
-					}
-				}
-			});
-		}});
-//DRAG_ALL+RIGHT+NORMAL
-		states.put(DRAG_ALL+RIGHT+NORMAL,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.RIGHT)){
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.RIGHT)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-		}});
-//TO_CREATE
-		states.put(TO_CREATE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.CREATE+Cursor.NORMAL);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Modulo modHover=getModuloHover(mouse);
-						if(modHover!=null){
-							tree.select(modHover);
-							setState(WAITING2+TO_CREATE);
-						}else{
-							final Conexao coxHover=getConexaoHover(mouse);
-							if(coxHover!=null){
-								final Segmento segHover=coxHover.segmentContains(mouse);
-								setState(WAITING2+TO_CREATE,segHover);
-							}else{
-								setState(WAITING1+TO_CREATE);
-							}
-						}
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+TO_CREATE);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseMovedAction(new Action(){
-				public void run(MouseEvent m){
-					tree.unSelectAll();
-					final Point mouse=getPosition(m.getPoint());
-					final Modulo modHover=getModuloHover(mouse);
-					if(modHover!=null){
-						tree.selectToBeCreator(modHover);
-					}else{
-						final Conexao coxHover=getConexaoHover(mouse);
-						if(coxHover!=null){
-							final Segmento segHover=coxHover.segmentContains(mouse);
-							tree.selectToBeCreator(segHover);
-						}
-					}
-					tree.draw();
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_UNSELECT);
-					setState(NORMAL);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE_SON);
-					setState(TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_DELETE);
-					setState(TO_DELETE);
-				}
-			});
-		}});
-//DRAG_ALL+TO_CREATE
-		states.put(DRAG_ALL+TO_CREATE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.CREATE+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(TO_CREATE);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_UNSELECT);
-					setState(DRAG_ALL+NORMAL);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE_SON);
-					setState(DRAG_ALL+TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_DELETE);
-					setState(DRAG_ALL+TO_DELETE);
-				}
-			});
-		}});
-//WAITING1+TO_CREATE
-		states.put(WAITING1+TO_CREATE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					if(getModulo()!=null){
-						tree.getCursor().set(Cursor.MOVE+Cursor.SELECT);
-					}else{
-						tree.getCursor().set(Cursor.CREATE+Cursor.SELECT);
-					}
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+WAITING1+TO_CREATE,getModulo());
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						addAndSelect(getModulo());
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(getModulo()!=null){
-							setState(MOVE+TO_CREATE,getModulo());
-						}else{
-							createMod(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-							setState(TO_CREATE);
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						if(getModulo()!=null){
-							tree.unSelectAll();
-							tree.draw();
-						}else{
-							createMod(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-						}
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(getModulo()!=null){
-							getModulo().setLocationIndex(mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-							setArea(Selecao.State.TO_CREATE_PAI,true,
-									mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-							setState(SELECT_AREA+TO_CREATE,getModulo());
-						}else{
-							setArea(Selecao.State.TO_CREATE,true,
-									mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-							setState(SELECT_AREA+TO_CREATE);
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						if(getModulo()!=null){
-							addAndSelect(getModulo());
-							setArea(Selecao.State.TO_SELECT,false,
-									mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-							setState(SELECT_AREA+NORMAL);
-						}else{
-							setArea(Selecao.State.TO_CREATE,false,
-									mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-							setState(SELECT_AREA+TO_CREATE);
-						}
-					}
-				}
-			});
-		}});
-//DRAG_ALL+WAITING1+TO_CREATE
-		states.put(DRAG_ALL+WAITING1+TO_CREATE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					if(getModulo()!=null){
-						tree.getCursor().set(Cursor.DRAG+Cursor.MOVE+Cursor.SELECT);
-					}else{
-						tree.getCursor().set(Cursor.DRAG+Cursor.CREATE+Cursor.SELECT);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(getModulo()!=null){
-							setState(DRAG_ALL+MOVE+TO_CREATE,getModulo());
-						}else{
-							createMod(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-							setState(DRAG_ALL+TO_CREATE);
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						if(getModulo()!=null){
-							tree.unSelectAll();
-							tree.draw();
-						}else{
-							createMod(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-						}
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(WAITING1+TO_CREATE,getModulo());
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-		}});
-//SELECT_AREA+TO_CREATE
-		states.put(SELECT_AREA+TO_CREATE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					if(getModulo()!=null){
-						tree.getCursor().set(Cursor.AREA_PAI+Cursor.SELECT);
-					}else{
-						tree.getCursor().set(Cursor.AREA_CREATE+Cursor.SELECT);
-					}
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+SELECT_AREA+TO_CREATE,getModulo());
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						tree.getSelecao().setEmpty();
-						addAndSelect(getModulo());
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						tree.getSelecao().setEmpty();
-						if(getModulo()!=null){
-							relateMods(tree.getSelectedObjetos().getModulos(),getModulo());
-							tree.unSelectAll();
-							tree.draw();
-							setState(MOVE+TO_CREATE,getModulo());
-						}else{
-							if(tree.getSelectedObjetos().getModulos().isEmpty()){
-								tree.draw();
-								setState(TO_CREATE);
-							}else{
-								final Modulo newMod=createModRelacionado(
-										mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY(),
-										tree.getSelectedObjetos().getModulos());
-								setState(MOVE+TO_CREATE,newMod);
-							}
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						tree.getSelecao().setEmpty();
-						if(getModulo()!=null){
-							addAndSelect(getModulo());
-							setArea(Selecao.State.TO_SELECT,false);
-							tree.getSelecao().setEmpty();
-							tree.draw();
-							setState(NORMAL);
-						}else{
-							if(tree.getSelectedObjetos().getModulos().isEmpty()){
-								tree.draw();
-								setState(NORMAL);
-							}else{
-								final Modulo newMod=createModRelacionado(
-										mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY(),
-										tree.getSelectedObjetos().getModulos());
-								setState(MOVE+TO_CREATE,newMod);
-							}
-						}
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(getModulo()!=null){
-							getModulo().setLocationIndex(mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-							setArea(Selecao.State.TO_CREATE_PAI,true,
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}else{
-							setArea(Selecao.State.TO_CREATE,true,
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						if(getModulo()!=null){
-							addAndSelect(getModulo());
-							setArea(Selecao.State.TO_SELECT,false,
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-							setState(SELECT_AREA+NORMAL);
-						}else{
-							setArea(Selecao.State.TO_CREATE,false,
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}
-					}
-				}
-			});
-			setKeyReleasedAction(new Action(){
-				public void run(KeyEvent k){
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_CONTROL:
-							if(getModulo()!=null){
-								addAndSelect(getModulo());
-								setArea(Selecao.State.TO_SELECT,false);
-								setState(SELECT_AREA+NORMAL);
-							}else{
-								setArea(Selecao.State.TO_CREATE,false);
-							}
-						break;
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					addAndSelect(getModulo());
-					setArea(Selecao.State.TO_SELECT,false);
-					setState(SELECT_AREA+NORMAL);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					addAndSelect(getModulo());
-					setArea(Selecao.State.TO_CREATE_SON,false);
-					setState(SELECT_AREA+TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					addAndSelect(getModulo());
-					setArea(Selecao.State.TO_DELETE,false);
-					setState(SELECT_AREA+TO_DELETE);
-				}
-			});
-		}});
-//DRAG_ALL+SELECT_AREA+TO_CREATE
-		states.put(DRAG_ALL+SELECT_AREA+TO_CREATE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					if(getModulo()!=null){
-						tree.getCursor().set(Cursor.DRAG+Cursor.AREA_PAI+Cursor.SELECT);
-					}else{
-						tree.getCursor().set(Cursor.DRAG+Cursor.AREA_CREATE+Cursor.SELECT);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						tree.getSelecao().setEmpty();
-						if(getModulo()!=null){
-							relateMods(tree.getSelectedObjetos().getModulos(),getModulo());
-							tree.unSelectAll();
-							tree.draw();
-							setState(DRAG_ALL+MOVE+TO_CREATE,getModulo());
-						}else{
-							if(tree.getSelectedObjetos().getModulos().isEmpty()){
-								tree.draw();
-								setState(DRAG_ALL+TO_CREATE);
-							}else{
-								final Modulo newMod=createModRelacionado(
-										mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY(),
-										tree.getSelectedObjetos().getModulos());
-								setState(DRAG_ALL+MOVE+TO_CREATE,newMod);
-							}
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						tree.getSelecao().setEmpty();
-						if(getModulo()!=null){
-							addAndSelect(getModulo());
-							setArea(Selecao.State.TO_CREATE,false);
-							tree.getSelecao().setEmpty();
-							tree.draw();
-							setState(DRAG_ALL+NORMAL);
-						}else{
-							if(tree.getSelectedObjetos().getModulos().isEmpty()){
-								tree.draw();
-								setState(DRAG_ALL+NORMAL);
-							}else{
-								final Modulo newMod=createModRelacionado(
-										mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY(),
-										tree.getSelectedObjetos().getModulos());
-								setState(DRAG_ALL+MOVE+TO_CREATE,newMod);
-							}
-						}
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(SELECT_AREA+TO_CREATE,getModulo());
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-			setKeyReleasedAction(new Action(){
-				public void run(KeyEvent k){
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_CONTROL:
-							if(getModulo()!=null){
-								addAndSelect(getModulo());
-								setArea(Selecao.State.TO_SELECT,false);
-								setState(DRAG_ALL+SELECT_AREA+NORMAL);
-							}else{
-								setArea(Selecao.State.TO_CREATE,false);
-							}
-						break;
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					addAndSelect(getModulo());
-					setArea(Selecao.State.TO_SELECT,false);
-					setState(DRAG_ALL+SELECT_AREA+NORMAL);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					addAndSelect(getModulo());
-					setArea(Selecao.State.TO_CREATE_SON,false);
-					setState(DRAG_ALL+SELECT_AREA+TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					addAndSelect(getModulo());
-					setArea(Selecao.State.TO_DELETE,false);
-					setState(DRAG_ALL+SELECT_AREA+TO_DELETE);
-				}
-			});
-		}});
-//WAITING2+TO_CREATE
-		states.put(WAITING2+TO_CREATE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					if(getModulo()!=null){
-						tree.getCursor().set(Cursor.MOVE+Cursor.SELECT);
-					}else{
-						tree.getCursor().set(Cursor.CREATE+Cursor.SELECT);
-					}
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+WAITING2+TO_CREATE,getModulo(),getSegmento());
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						addAndSelect(getModulo());
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(getModulo()!=null){
-							relateMods(tree.getSelectedObjetos().getModulos(),getModulo());
-							tree.draw();
-							setState(MOVE+TO_CREATE,getModulo());
-						}else{
-							if(getSegmento()!=null){
-								createNod(getSegmento(),mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY(),true);
-								setState(TO_CREATE);
-							}else{
-								final Modulo newMod=createModRelacionado(
-										mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY(),
-										tree.getSelectedObjetos().getModulos());
-								setState(MOVE+TO_CREATE,newMod);
-							}
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						if(getModulo()!=null){
-							addAndSelect(getModulo());
-							setState(NORMAL);
-						}else{
-							if(getSegmento()!=null){
-								createNod(getSegmento(),mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY(),true);
-								setState(NORMAL);
-							}else{
-								final Modulo newMod=createModRelacionado(
-										mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY(),
-										tree.getSelectedObjetos().getModulos());
-								setState(MOVE+TO_CREATE,newMod);
-							}
-						}
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT)){
-						if(getModulo()!=null){
-							getModulo().setLocationIndex(mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-							relateMods(tree.getSelectedObjetos().getModulos(),getModulo());
-							setState(DRAG+TO_CREATE,getModulo());
-						}else{
-							if(getSegmento()!=null){
-								final Nodulo newNod=createNod(getSegmento(),mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY(),false);
-								tree.unSelect(newNod);
-								setState(DRAG+TO_CREATE,newNod);
-							}else{
-								final Modulo newMod=createModRelacionado(
-										mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY(),
-										tree.getSelectedObjetos().getModulos());
-								setState(DRAG+TO_CREATE,newMod);
-							}
-						}
-					}
-				}
-			});
-		}});
-//DRAG_ALL+WAITING2+TO_CREATE
-		states.put(DRAG_ALL+WAITING2+TO_CREATE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					if(getModulo()!=null){
-						tree.getCursor().set(Cursor.DRAG+Cursor.MOVE+Cursor.SELECT);
-					}else{
-						tree.getCursor().set(Cursor.DRAG+Cursor.CREATE+Cursor.SELECT);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(getModulo()!=null){
-							relateMods(tree.getSelectedObjetos().getModulos(),getModulo());
-							tree.draw();
-							setState(DRAG_ALL+MOVE+TO_CREATE,getModulo());
-						}else{
-							if(getSegmento()!=null){
-								createNod(getSegmento(),mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY(),true);
-								setState(DRAG_ALL+TO_CREATE);
-							}else{
-								final Modulo newMod=createModRelacionado(
-										mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY(),
-										tree.getSelectedObjetos().getModulos());
-								setState(DRAG_ALL+MOVE+TO_CREATE,newMod);
-							}
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						if(getModulo()!=null){
-							addAndSelect(getModulo());
-							setState(DRAG_ALL+NORMAL);
-						}else{
-							if(getSegmento()!=null){
-								createNod(getSegmento(),mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY(),true);
-								setState(DRAG_ALL+NORMAL);
-							}else{
-								final Modulo newMod=createModRelacionado(
-										mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY(),
-										tree.getSelectedObjetos().getModulos());
-								setState(DRAG_ALL+MOVE+TO_CREATE,newMod);
-							}
-						}
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(WAITING2+TO_CREATE,getModulo(),getSegmento());
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-		}});
-//DRAG+TO_CREATE
-		states.put(DRAG+TO_CREATE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.MOVE+Cursor.SELECT);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+DRAG+TO_CREATE,getModulo(),getNodulo());
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						if(getModulo()!=null){
-							addAndSelect(getModulo());
-						}else if(getNodulo()!=null){
-							addAndSelect(getNodulo());
-						}
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(getModulo()!=null){
-							setState(MOVE+TO_CREATE,getModulo());
-						}else{
-							addAndSelect(getNodulo());
-							setState(NORMAL);
-						}
-					
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						if(getModulo()!=null){
-							addAndSelect(getModulo());
-						}else if(getNodulo()!=null){
-							addAndSelect(getNodulo());
-						}
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT)){
-						if(getModulo()!=null){
-							getModulo().setLocationIndex(mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}else if(getNodulo()!=null){
-							getNodulo().setLocationIndex(mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}
-						tree.draw();
-					}
-				}
-			});
-		}});
-//DRAG_ALL+DRAG+TO_CREATE
-		states.put(DRAG_ALL+DRAG+TO_CREATE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.MOVE+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(getModulo()!=null){
-							setState(DRAG_ALL+MOVE+TO_CREATE,getModulo());
-						}else{
-							addAndSelect(getNodulo());
-							setState(DRAG_ALL+NORMAL);
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						if(getModulo()!=null){
-							addAndSelect(getModulo());
-						}else if(getNodulo()!=null){
-							addAndSelect(getNodulo());
-						}
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG+TO_CREATE,getModulo(),getNodulo());
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-		}});
-//MOVE+TO_CREATE
-		states.put(MOVE+TO_CREATE,new State(){
-			private boolean lock=false;
-			{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.MOVE+Cursor.NORMAL);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Modulo modHover=getModuloHover(mouse,getModulo());
-						if(modHover!=null){
-							tree.select(modHover);
-							setState(WAITING2+TO_CREATE,getModulo());
-						}else{
-							setState(WAITING1+TO_CREATE,getModulo());
-						}
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+MOVE+TO_CREATE,getModulo());
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						addAndSelect(getModulo());
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						//NADA
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						addAndSelect(getModulo());
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseMovedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseMovedAtual=getGridPosition(m.getPoint());
-					final Point mouse=getPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.CTRL)){
-						getModulo().setLocationIndex(mouseMovedAtual.x-Tree.getLocalX(),mouseMovedAtual.y-Tree.getLocalY());
-						tree.unSelectAll();
-						final Modulo modHover=getModuloHover(mouse,getModulo());
-						if(modHover!=null){
-							tree.selectToBePai(modHover);
-						}
-						tree.draw();
-					}else{
-						getModulo().setLocationIndex(mouseMovedAtual.x-Tree.getLocalX(),mouseMovedAtual.y-Tree.getLocalY());
-						tree.draw();
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT)){
-						addAndSelect(getModulo());
-						setArea(Selecao.State.TO_SELECT,false,
-								mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-								mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						setState(SELECT_AREA+NORMAL);
-					}
-				}
-			});
-			setKeyPressedAction(new Action(){
-				public void run(KeyEvent k){
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_CONTROL:
-							if(lock)return;
-							lock=true;
-							final Modulo modHover=getModuloHover(getNonGridPosition(mouseMoved),getModulo());
-							if(modHover!=null){
-								tree.selectToBePai(modHover);
-								tree.draw();
-							}
-							tree.getCursor().set(Cursor.PAI+Cursor.NORMAL);
-						break;
-					}
-				}
-			});
-			setKeyReleasedAction(new Action(){
-				public void run(KeyEvent k){
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_CONTROL:
-							lock=false;
-							tree.unSelectAllMods();
-							tree.draw();
-							tree.getCursor().set(Cursor.MOVE+Cursor.NORMAL);
-						break;
-					}
-				}
-			});
-		}});
-//DRAG_ALL+MOVE+TO_CREATE
-		states.put(DRAG_ALL+MOVE+TO_CREATE,new State(){
-			private boolean lock=false;
-			{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.MOVE+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						//NADA
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						addAndSelect(getModulo());
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(MOVE+TO_CREATE,getModulo());
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-			setKeyPressedAction(new Action(){
-				public void run(KeyEvent k){
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_CONTROL:
-							if(lock)return;
-							lock=true;
-							final Modulo modHover=getModuloHover(getNonGridPosition(mouseMoved),getModulo());
-							if(modHover!=null){
-								tree.selectToBePai(modHover);
-								tree.draw();
-							}
-							tree.getCursor().set(Cursor.DRAG+Cursor.PAI+Cursor.NORMAL);
-						break;
-					}
-				}
-			});
-			setKeyReleasedAction(new Action(){
-				public void run(KeyEvent k){
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_CONTROL:
-							lock=false;
-							tree.unSelectAllMods();
-							tree.draw();
-							tree.getCursor().set(Cursor.DRAG+Cursor.MOVE+Cursor.NORMAL);
-						break;
-					}
-				}
-			});
-		}});
-//TO_CONNECT
-		states.put(TO_CONNECT,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.SON+Cursor.NORMAL);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						Modulo modSelec=getModuloHover(mouse);
-						if(modSelec==Tree.getMestre())modSelec=null;
-						if(modSelec!=null){
-							setState(WAITING2+TO_CONNECT,modSelec);
-						}else{
-							setState(WAITING1+TO_CONNECT);
-						}
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+TO_CONNECT);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseMovedAction(new Action(){
-				public void run(MouseEvent m){
-					tree.unSelectAllMods();
-					final Point mouse=getPosition(m.getPoint());
-					final Modulo modHover=getModuloHover(mouse);
-					tree.selectToBeSon(modHover);
-					tree.draw();
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE);
-					setState(TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_UNSELECT);
-					setState(NORMAL);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_DELETE);
-					setState(TO_DELETE);
-				}
-			});
-		}});
-//DRAG_ALL+TO_CONNECT
-		states.put(DRAG_ALL+TO_CONNECT,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.SON+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(TO_CONNECT);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE);
-					setState(DRAG_ALL+TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_UNSELECT);
-					setState(DRAG_ALL+NORMAL);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_DELETE);
-					setState(DRAG_ALL+TO_DELETE);
-				}
-			});
-		}});
-//WAITING1+TO_CONNECT
-		states.put(WAITING1+TO_CONNECT,new State(){{
-			setAction(new Action(){
-				public void run(){
-					if(Tree.getGhost().getConexoes().isEmpty()){
-						tree.getCursor().set(Cursor.SON+Cursor.SELECT);
-					}else{
-						tree.getCursor().set(Cursor.PAI+Cursor.SELECT);
-					}
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+WAITING1+TO_CONNECT);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						deleteGhostCoxs();
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(Tree.getGhost().getConexoes().isEmpty()){
-							setState(TO_CONNECT);
-						}else{
-							createNodOnGhostCoxs(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-							setState(MOVE+TO_CONNECT);
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						deleteGhostCoxs();
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						Tree.getGhost().setLocationIndex(mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						if(Tree.getGhost().getConexoes().isEmpty()){
-							setArea(Selecao.State.TO_CREATE_SON,true,
-									mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}else{
-							setArea(Selecao.State.TO_CREATE_PAI,true,
-									mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}
-						tree.draw();
-						setState(SELECT_AREA+TO_CONNECT);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						Tree.getGhost().setLocationIndex(mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						if(Tree.getGhost().getConexoes().isEmpty()){
-							setArea(Selecao.State.TO_CREATE_SON,false,
-									mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}else{
-							setArea(Selecao.State.TO_CREATE_PAI,false,
-									mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}
-						tree.draw();
-						setState(SELECT_AREA+TO_CONNECT);
-					}
-				}
-			});
-		}});
-//DRAG_ALL+WAITING1+TO_CONNECT
-		states.put(DRAG_ALL+WAITING1+TO_CONNECT,new State(){{
-			setAction(new Action(){
-				public void run(){
-					if(Tree.getGhost().getConexoes().isEmpty()){
-						tree.getCursor().set(Cursor.DRAG+Cursor.SON+Cursor.SELECT);
-					}else{
-						tree.getCursor().set(Cursor.DRAG+Cursor.PAI+Cursor.SELECT);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(Tree.getGhost().getConexoes().isEmpty()){
-							setState(DRAG_ALL+TO_CONNECT);
-						}else{
-							createNodOnGhostCoxs(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-							setState(DRAG_ALL+MOVE+TO_CONNECT);
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						deleteGhostCoxs();
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						setState(WAITING1+TO_CONNECT);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-		}});
-//SELECT_AREA+TO_CONNECT
-		states.put(SELECT_AREA+TO_CONNECT,new State(){{
-			setAction(new Action(){
-				public void run(){
-					if(Tree.getGhost().getConexoes().isEmpty()){
-						tree.getCursor().set(Cursor.AREA_SON+Cursor.SELECT);
-					}else{
-						tree.getCursor().set(Cursor.AREA_PAI+Cursor.SELECT);
-					}
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+SELECT_AREA+TO_CONNECT);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						deleteGhostCoxs();
-						tree.getSelecao().setEmpty();
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						tree.getSelecao().setEmpty();
-						if(tree.getSelectedObjetos().getModulos().isEmpty()){
-							tree.draw();
-							if(Tree.getGhost().getConexoes().isEmpty()){
-								setState(TO_CONNECT);
-							}else{
-								setState(MOVE+TO_CONNECT);
-							}
-						}else{
-							if(Tree.getGhost().getConexoes().isEmpty()){
-								Tree.getGhost().setLocationIndex(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-								relateToGhost(tree.getSelectedObjetos().getModulos());
-							}else{
-								relateToMods(tree.getSelectedObjetos().getModulos());
-							}
-							setState(MOVE+TO_CONNECT);
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						tree.getSelecao().setEmpty();
-						if(tree.getSelectedObjetos().getModulos().isEmpty()){
-							deleteGhostCoxs();
-							setState(NORMAL);
-						}else{
-							if(Tree.getGhost().getConexoes().isEmpty()){
-								Tree.getGhost().setLocationIndex(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-								relateToGhost(tree.getSelectedObjetos().getModulos());
-								setState(MOVE+TO_CONNECT);
-							}else{
-								relateToMods(tree.getSelectedObjetos().getModulos());
-								deleteGhostCoxs();
-								setState(NORMAL);
-							}
-						}
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(Tree.getGhost().getConexoes().isEmpty()){
-							setArea(Selecao.State.TO_CREATE_SON,true,
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}else{
-							Tree.getGhost().setLocationIndex(mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-							setArea(Selecao.State.TO_CREATE_PAI,true,
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						if(Tree.getGhost().getConexoes().isEmpty()){
-							setArea(Selecao.State.TO_CREATE_SON,false,
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}else{
-							Tree.getGhost().setLocationIndex(mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-							setArea(Selecao.State.TO_CREATE_PAI,false,
-									mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						}
-					}
-				}
-			});
-			setKeyReleasedAction(new Action(){
-				public void run(KeyEvent k){
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_CONTROL:
-							if(Tree.getGhost().getConexoes().isEmpty()){
-								setArea(Selecao.State.TO_CREATE_SON,false);
-							}else setArea(Selecao.State.TO_CREATE_PAI,false);
-						break;
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					deleteGhostCoxs();
-					setArea(Selecao.State.TO_CREATE,false);
-					setState(SELECT_AREA+TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					deleteGhostCoxs();
-					setArea(Selecao.State.TO_SELECT,false);
-					setState(SELECT_AREA+NORMAL);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					deleteGhostCoxs();
-					setArea(Selecao.State.TO_DELETE,false);
-					setState(SELECT_AREA+TO_DELETE);
-				}
-			});
-		}});
-//DRAG_ALL+SELECT_AREA+TO_CONNECT
-		states.put(DRAG_ALL+SELECT_AREA+TO_CONNECT,new State(){{
-			setAction(new Action(){
-				public void run(){
-					if(Tree.getGhost().getConexoes().isEmpty()){
-						tree.getCursor().set(Cursor.DRAG+Cursor.AREA_SON+Cursor.SELECT);
-					}else{
-						tree.getCursor().set(Cursor.DRAG+Cursor.AREA_PAI+Cursor.SELECT);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						tree.getSelecao().setEmpty();
-						if(tree.getSelectedObjetos().getModulos().isEmpty()){
-							tree.draw();
-							if(Tree.getGhost().getConexoes().isEmpty()){
-								setState(DRAG_ALL+TO_CONNECT);
-							}else{
-								setState(DRAG_ALL+MOVE+TO_CONNECT);
-							}
-						}else{
-							if(Tree.getGhost().getConexoes().isEmpty()){
-								Tree.getGhost().setLocationIndex(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-								relateToGhost(tree.getSelectedObjetos().getModulos());
-							}else{
-								relateToMods(tree.getSelectedObjetos().getModulos());
-							}
-							setState(DRAG_ALL+MOVE+TO_CONNECT);
-						}
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						tree.getSelecao().setEmpty();
-						if(tree.getSelectedObjetos().getModulos().isEmpty()){
-							deleteGhostCoxs();
-							setState(DRAG_ALL+NORMAL);
-						}else{
-							if(Tree.getGhost().getConexoes().isEmpty()){
-								Tree.getGhost().setLocationIndex(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-								relateToGhost(tree.getSelectedObjetos().getModulos());
-								setState(DRAG_ALL+MOVE+TO_CONNECT);
-							}else{
-								relateToMods(tree.getSelectedObjetos().getModulos());
-								deleteGhostCoxs();
-								setState(DRAG_ALL+NORMAL);
-							}
-						}
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(SELECT_AREA+TO_CREATE);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-			setKeyReleasedAction(new Action(){
-				public void run(KeyEvent k){
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_CONTROL:
-							if(Tree.getGhost().getConexoes().isEmpty()){
-								setArea(Selecao.State.TO_CREATE_SON,false);
-							}else setArea(Selecao.State.TO_CREATE_PAI,false);
-						break;
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					deleteGhostCoxs();
-					setArea(Selecao.State.TO_CREATE,false);
-					setState(DRAG_ALL+SELECT_AREA+TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					deleteGhostCoxs();
-					setArea(Selecao.State.TO_SELECT,false);
-					setState(DRAG_ALL+SELECT_AREA+NORMAL);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					deleteGhostCoxs();
-					setArea(Selecao.State.TO_DELETE,false);
-					setState(DRAG_ALL+SELECT_AREA+TO_DELETE);
-				}
-			});
-		}});
-//WAITING2+TO_CONNECT
-		states.put(WAITING2+TO_CONNECT,new State(){{
-			setAction(new Action(){
-				public void run(){
-					if(Tree.getGhost().getConexoes().isEmpty()){
-						tree.getCursor().set(Cursor.SON+Cursor.SELECT);
-					}else{
-						tree.getCursor().set(Cursor.PAI+Cursor.SELECT);
-					}
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+WAITING2+TO_CONNECT,getModulo());
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						deleteGhostCoxs();
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(Tree.getGhost().getConexoes().isEmpty()){
-							Tree.getGhost().setLocationIndex(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-							relateToGhost(getModulo());
-						}else{
-							relateToMods(getModulo());
-						}
-						setState(MOVE+TO_CONNECT);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						if(Tree.getGhost().getConexoes().isEmpty()){
-							Tree.getGhost().setLocationIndex(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-							relateToGhost(getModulo());
-							setState(MOVE+TO_CONNECT);
-						}else{
-							relateToMods(getModulo());
-							deleteGhostCoxs();
-							setState(NORMAL);
-						}
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT)){
-						Tree.getGhost().setLocationIndex(mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						relateToGhost(getModulo());
-						setState(DRAG+TO_CONNECT);
-					}
-				}
-			});
-		}});
-//DRAG_ALL+WAITING2+TO_CONNECT
-		states.put(DRAG_ALL+WAITING2+TO_CONNECT,new State(){{
-			setAction(new Action(){
-				public void run(){
-					if(Tree.getGhost().getConexoes().isEmpty()){
-						tree.getCursor().set(Cursor.DRAG+Cursor.SON+Cursor.SELECT);
-					}else{
-						tree.getCursor().set(Cursor.DRAG+Cursor.PAI+Cursor.SELECT);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						if(Tree.getGhost().getConexoes().isEmpty()){
-							Tree.getGhost().setLocationIndex(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-							relateToGhost(getModulo());
-						}else{
-							relateToMods(getModulo());
-						}
-						setState(DRAG_ALL+MOVE+TO_CONNECT);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						if(Tree.getGhost().getConexoes().isEmpty()){
-							Tree.getGhost().setLocationIndex(mouseReleased.x-Tree.getLocalX(),mouseReleased.y-Tree.getLocalY());
-							relateToGhost(getModulo());
-							setState(DRAG_ALL+MOVE+TO_CONNECT);
-						}else{
-							relateToMods(getModulo());
-							deleteGhostCoxs();
-							setState(DRAG_ALL+NORMAL);
-						}
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(WAITING2+TO_CONNECT,getModulo());
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-		}});
-//DRAG+TO_CONNECT
-		states.put(DRAG+TO_CONNECT,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.MOVE+Cursor.SELECT);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+DRAG+TO_CONNECT);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						deleteGhostCoxs();
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Modulo modSelec=getModuloHover(mouse);
-						if(modSelec!=null){
-							relateToMods(modSelec);
-						}
-						setState(MOVE+TO_CONNECT);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Modulo modSelec=getModuloHover(mouse);
-						if(modSelec!=null){
-							relateToMods(modSelec);
-						}
-						deleteGhostCoxs();
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT)){
-						Tree.getGhost().setLocationIndex(mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						tree.unSelectAll();
-						final Point mouse=getPosition(m.getPoint());
-						final Modulo modHover=getModuloHover(mouse);
-						if(modHover!=null){
-							tree.selectToBePai(modHover);
-						}
-						tree.draw();
-					}
-				}
-			});
-		}});
-//DRAG_ALL+DRAG+TO_CONNECT
-		states.put(DRAG_ALL+DRAG+TO_CONNECT,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.MOVE+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Modulo modSelec=getModuloHover(mouse);
-						if(modSelec!=null){
-							relateToMods(modSelec);
-						}
-						setState(DRAG_ALL+MOVE+TO_CONNECT);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Modulo modSelec=getModuloHover(mouse);
-						if(modSelec!=null){
-							relateToMods(modSelec);
-						}
-						deleteGhostCoxs();
-						tree.draw();
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG+TO_CONNECT);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-		}});
-//MOVE+TO_CONNECT
-		states.put(MOVE+TO_CONNECT,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.PAI+Cursor.NORMAL);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Modulo modSelec=getModuloHover(mouse);
-						if(modSelec!=null){
-							setState(WAITING2+TO_CONNECT,modSelec);
-						}else{
-							setState(WAITING1+TO_CONNECT);
-						}
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+MOVE+TO_CONNECT);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						deleteGhostCoxs();
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseMovedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseMovedAtual=getGridPosition(m.getPoint());
-					Tree.getGhost().setLocationIndex(mouseMovedAtual.x-Tree.getLocalX(),mouseMovedAtual.y-Tree.getLocalY());
-					tree.unSelectAllMods();
-					final Point mouse=getPosition(m.getPoint());
-					final Modulo modHover=getModuloHover(mouse);
-					if(modHover!=null){
-						tree.selectToBePai(modHover);
-					}
-					tree.draw();
-				}
-			});
-		}});
-//DRAG_ALL+MOVE+TO_CONNECT
-		states.put(DRAG_ALL+MOVE+TO_CONNECT,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.PAI+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(MOVE+TO_CONNECT);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-		}});
-//TO_DELETE
-		states.put(TO_DELETE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DELETE+Cursor.NORMAL);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						setState(WAITING1+TO_DELETE);
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+TO_DELETE);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseMovedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouse=getPosition(m.getPoint());
-					tree.unSelectAll();
-					final Modulo modDel=getModuloHover(mouse);
-					if(modDel!=null){
-						tree.selectToBeDeleted(modDel);
-					}else{
-						final Nodulo nodDel=getNoduloHover(mouse);
-						if(nodDel!=null){
-							tree.selectToBeDeleted(nodDel);
-						}else{
-							final Conexao coxDel=getConexaoHover(mouse);
-							if(coxDel!=null){
-								tree.selectToBeDeleted(coxDel);
-							}
-						}
-					}
-					tree.draw();
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE);
-					setState(TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE_SON);
-					setState(TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_UNSELECT);
-					setState(NORMAL);
-				}
-			});
-		}});
-//DRAG_ALL+TO_DELETE
-		states.put(DRAG_ALL+TO_DELETE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.DELETE+Cursor.SELECT);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.LEFT)){
-						setState(DRAG_ALL+WAITING1+TO_DELETE);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(TO_DELETE);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE);
-					setState(DRAG_ALL+TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_CREATE_SON);
-					setState(DRAG_ALL+TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setHover(Selecao.State.TO_UNSELECT);
-					setState(DRAG_ALL+NORMAL);
-				}
-			});
-		}});
-//WAITING1+TO_DELETE
-		states.put(WAITING1+TO_DELETE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DELETE+Cursor.SELECT);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+WAITING1+TO_DELETE);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Objeto objSelec=getObjetoHover(mouse);
-						if(objSelec!=null){
-							if(tree.getSelectedObjetos().getAll().containsValue(objSelec)){
-								delUndoable(new ArrayList<>(tree.getSelectedObjetos().getAll().values()));
-							}
-						}
-						tree.draw();
-						setState(TO_DELETE);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Objeto objSelec=getObjetoHover(mouse);
-						if(objSelec!=null){
-							if(tree.getSelectedObjetos().getAll().containsValue(objSelec)){
-								delUndoable(new ArrayList<>(tree.getSelectedObjetos().getAll().values()));
-							}
-						}
-						tree.draw();
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						setArea(Selecao.State.TO_DELETE,true,
-								mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-								mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						setState(SELECT_AREA+TO_DELETE);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						setArea(Selecao.State.TO_DELETE,false,
-								mouseDragged.x-Tree.getLocalX(),mouseDragged.y-Tree.getLocalY(),
-								mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						setState(SELECT_AREA+TO_DELETE);
-					}
-				}
-			});
-		}});
-//DRAG_ALL+WAITING1+TO_DELETE
-		states.put(DRAG_ALL+WAITING1+TO_DELETE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.DELETE+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Objeto objSelec=getObjetoHover(mouse);
-						if(objSelec!=null){
-							if(tree.getSelectedObjetos().getAll().containsValue(objSelec)){
-								delUndoable(new ArrayList<>(tree.getSelectedObjetos().getAll().values()));
-							}
-						}
-						tree.draw();
-						setState(DRAG_ALL+TO_DELETE);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						final Point mouse=getPosition(m.getPoint());
-						final Objeto objSelec=getObjetoHover(mouse);
-						if(objSelec!=null){
-							if(tree.getSelectedObjetos().getAll().containsValue(objSelec)){
-								delUndoable(new ArrayList<>(tree.getSelectedObjetos().getAll().values()));
-							}
-						}
-						tree.draw();
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(WAITING1+TO_DELETE);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-		}});
-//SELECT_AREA+TO_DELETE
-		states.put(SELECT_AREA+TO_DELETE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.AREA_DELETE+Cursor.SELECT);
-				}
-			});
-			setMousePressedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(DRAG_ALL+SELECT_AREA+TO_DELETE);
-					}else if(Cursor.match(m,Cursor.RIGHT)){
-						tree.getSelecao().setEmpty();
-						tree.unSelectAll();
-						tree.draw();
-						setState(RIGHT+WAITING1+NORMAL);
-					}
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						delUndoable(new ArrayList<>(tree.getSelectedObjetos().getAll().values()));
-						tree.getSelecao().setEmpty();
-						tree.draw();
-						setState(TO_DELETE);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						delUndoable(new ArrayList<>(tree.getSelectedObjetos().getAll().values()));
-						tree.getSelecao().setEmpty();
-						tree.draw();
-						setState(NORMAL);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						setArea(Selecao.State.TO_DELETE,true,
-								mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-						tree.draw();
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						setArea(Selecao.State.TO_DELETE,false,
-								mouseDraggedAtual.x-Tree.getLocalX(),mouseDraggedAtual.y-Tree.getLocalY());
-					}
-				}
-			});
-			setKeyReleasedAction(new Action(){
-				public void run(KeyEvent k){
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_CONTROL:
-							setArea(Selecao.State.TO_DELETE,false);
-						break;
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setArea(Selecao.State.TO_CREATE,false);
-					setState(SELECT_AREA+TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setArea(Selecao.State.TO_CREATE_SON,false);
-					setState(SELECT_AREA+TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setArea(Selecao.State.TO_SELECT,false);
-					setState(SELECT_AREA+NORMAL);
-				}
-			});
-		}});
-//DRAG_ALL+SELECT_AREA+TO_DELETE
-		states.put(DRAG_ALL+SELECT_AREA+TO_DELETE,new State(){{
-			setAction(new Action(){
-				public void run(){
-					tree.getCursor().set(Cursor.DRAG+Cursor.AREA_DELETE+Cursor.SELECT);
-				}
-			});
-			setMouseReleasedAction(new Action(){
-				public void run(MouseEvent m){
-					if(Cursor.match(m,Cursor.CTRL,Cursor.LEFT)){
-						delUndoable(new ArrayList<>(tree.getSelectedObjetos().getAll().values()));
-						tree.getSelecao().setEmpty();
-						tree.draw();
-						setState(DRAG_ALL+TO_DELETE);
-					}else if(Cursor.match(m,Cursor.LEFT)){
-						delUndoable(new ArrayList<>(tree.getSelectedObjetos().getAll().values()));
-						tree.getSelecao().setEmpty();
-						tree.draw();
-						setState(DRAG_ALL+NORMAL);
-					}else if(Cursor.match(m,Cursor.MIDDLE)){
-						setState(SELECT_AREA+TO_DELETE);
-					}
-				}
-			});
-			setMouseDraggedAction(new Action(){
-				public void run(MouseEvent m){
-					final Point mouseDraggedAtual=getGridPosition(m.getPoint());
-					if(Cursor.match(m,Cursor.LEFT,Cursor.MIDDLE)){
-						Tree.setLocal(Tree.getLocalX()-(mouseDragged.x-mouseDraggedAtual.x),Tree.getLocalY()-(mouseDragged.y-mouseDraggedAtual.y));
-						tree.draw();
-					}
-				}
-			});
-			setKeyReleasedAction(new Action(){
-				public void run(KeyEvent k){
-					switch(k.getKeyCode()){
-						case KeyEvent.VK_CONTROL:
-							setArea(Selecao.State.TO_DELETE,false);
-						break;
-					}
-				}
-			});
-			setCreateAction(new Action(){
-				public void run(){
-					setArea(Selecao.State.TO_CREATE,false);
-					setState(DRAG_ALL+SELECT_AREA+TO_CREATE);
-				}
-			});
-			setConnectAction(new Action(){
-				public void run(){
-					setArea(Selecao.State.TO_CREATE_SON,false);
-					setState(DRAG_ALL+SELECT_AREA+TO_CONNECT);
-				}
-			});
-			setDeleteAction(new Action(){
-				public void run(){
-					setArea(Selecao.State.TO_SELECT,false);
-					setState(DRAG_ALL+SELECT_AREA+NORMAL);
-				}
-			});
-		}});
-		setState(NORMAL);
 	}
 //INSERIR AÇÕES
 	public void putIntputs(){
+		final JFrame janela=(JFrame)tree.getPainel().getJanela();
 		janela.addMouseWheelListener(new MouseWheelListener(){
 			public void mouseWheelMoved(MouseWheelEvent w){
 				if(!tree.getRelativeBounds().contains(getPosition(w.getPoint())))return;
@@ -3067,8 +68,8 @@ public class Actions{
 				mouseMoved.y-=diff.y;
 				tree.draw();
 				updateTituloBounds();
-				if(tree.getTitulo().getObjeto()!=null){		//ATUALIZA O TAMANHO DA FONTE DE TÍTULO
-					tree.updateTituloFont();
+				if(tree.getUI().getTitulo().getObjeto()!=null){		//ATUALIZA O TAMANHO DA FONTE DE TÍTULO
+					tree.getUI().updateTituloFont();
 				}
 			}
 		});
@@ -3079,7 +80,7 @@ public class Actions{
 				mousePressed=new Point(mouseAtual);
 				mouseDragged=new Point(mouseAtual);
 				mouseMoved=new Point(mouseAtual);
-				final State state=getStates().get(getState());
+				final TreeST.State state=tree.getStateContent(tree.getState());
 				if(state.getMousePressedAction()!=null)state.getMousePressedAction().run(m);
 			}
 			public void mouseReleased(MouseEvent m){
@@ -3088,7 +89,7 @@ public class Actions{
 				mouseReleased=new Point(mouseAtual);
 				mouseDragged=new Point(mouseAtual);
 				mouseMoved=new Point(mouseAtual);
-				final State state=getStates().get(getState());
+				final TreeST.State state=tree.getStateContent(tree.getState());
 				if(state.getMouseReleasedAction()!=null)state.getMouseReleasedAction().run(m);
 			}
 			public void mouseClicked(MouseEvent m){
@@ -3098,7 +99,7 @@ public class Actions{
 				mouseReleased=new Point(mouseAtual);
 				mouseDragged=new Point(mouseAtual);
 				mouseMoved=new Point(mouseAtual);
-				final State state=getStates().get(getState());
+				final TreeST.State state=tree.getStateContent(tree.getState());
 				if(state.getMouseCliquedAction()!=null)state.getMouseCliquedAction().run(m);
 			}
 		});
@@ -3107,7 +108,7 @@ public class Actions{
 				if(!tree.isEnabled())return;
 				final Point mouseDraggedAtual=getGridPosition(m.getPoint());
 				if(!mouseDraggedAtual.equals(mouseDragged)){
-					final State state=getStates().get(getState());
+					final TreeST.State state=tree.getStateContent(tree.getState());
 					if(state.getMouseDraggedAction()!=null)state.getMouseDraggedAction().run(m);
 					mouseDragged=new Point(mouseDraggedAtual);
 					mouseMoved=new Point(mouseDraggedAtual);
@@ -3117,7 +118,7 @@ public class Actions{
 				if(!tree.isEnabled())return;	//IGNORA SE DESFOCADO
 				final Point mouseMovedAtual=getGridPosition(m.getPoint());
 				if(!mouseMovedAtual.equals(mouseMoved)){
-					final State state=getStates().get(getState());
+					final TreeST.State state=tree.getStateContent(tree.getState());
 					if(state.getMouseMovedAction()!=null)state.getMouseMovedAction().run(m);
 					mouseMoved=new Point(mouseMovedAtual);
 				}
@@ -3126,12 +127,12 @@ public class Actions{
 		janela.addKeyListener(new KeyAdapter(){
 			public void keyPressed(KeyEvent k){
 				if(!tree.isEnabled())return;
-				final State state=getStates().get(getState());
+				final TreeST.State state=tree.getStateContent(tree.getState());
 				if(state.getKeyPressedAction()!=null)state.getKeyPressedAction().run(k);
 			}
 			public void keyReleased(KeyEvent k){
 				if(!tree.isEnabled())return;
-				final State state=getStates().get(getState());
+				final TreeST.State state=tree.getStateContent(tree.getState());
 				if(state.getKeyReleasedAction()!=null)state.getKeyReleasedAction().run(k);
 			}
 		});
@@ -3143,14 +144,105 @@ public class Actions{
 			public void windowClosed(WindowEvent w){}
 			public void windowActivated(WindowEvent w){
 				tree.setEnabled(true);
-				if(tree.getTitulo().isVisible())tree.getTitulo().requestFocus();
+				if(tree.getUI().getTitulo().isVisible())tree.getUI().getTitulo().requestFocus();
 			}
 			public void windowDeactivated(WindowEvent w){
 				if(!janela.isUndecorated())tree.setEnabled(false);
 			}
 		});
 	}
-	private void startAutoDrag(){
+	public void setFocusOn(Objeto[]objs){
+		if(objs==null)return;
+		int xIndexMin=0;
+		int yIndexMin=0;
+		int xIndexMax=0;
+		int yIndexMax=0;
+		boolean firstRun=true;
+		for(Objeto obj:objs){
+			if(obj==null)continue;
+			int x=0;
+			int y=0;
+			switch(obj.getTipo()){
+				case MODULO:
+					final Modulo mod=(Modulo)obj;
+					x=mod.getMeioXIndex();
+					y=mod.getMeioYIndex();
+				break;
+				case CONEXAO:
+					final Conexao cox=(Conexao)obj;
+					final Rectangle coxForm=cox.getFormIndex().getBounds();
+					x=coxForm.x+(coxForm.width/2);
+					y=coxForm.y+(coxForm.height/2);
+				break;
+				case NODULO:
+					final Nodulo nod=(Nodulo)obj;
+					x=nod.getXIndex();
+					y=nod.getYIndex();
+				break;
+				case SEGMENTO:break;
+			}
+			if(firstRun){
+				xIndexMin=x;
+				yIndexMin=y;
+				xIndexMax=x;
+				yIndexMax=y;
+				firstRun=false;
+			}else{
+				xIndexMin=Math.min(xIndexMin,x);
+				yIndexMin=Math.min(yIndexMin,y);
+				xIndexMax=Math.max(xIndexMax,x);
+				yIndexMax=Math.max(yIndexMax,y);
+			}
+		}
+		final int meioX=(xIndexMax+xIndexMin)/2;
+		final int meioY=(yIndexMax+yIndexMin)/2;
+		final Point telaMeio=getGridPosition(new Point(tree.getPainel().getJanela().getWidth()/2,tree.getPainel().getJanela().getHeight()/2));
+		Tree.setLocal(telaMeio.x-meioX,telaMeio.y-meioY);	//ALINHA O MEIO DA JANELA COM O MEIO DOS OBJS FOCADOS
+	}
+	private Runnable animation;
+	public void animate(Point localIni,Point localFim){
+		new Thread(animation=new Runnable(){
+			public void run(){
+				final double diffX=localFim.x-localIni.x;
+				final double diffY=localFim.y-localIni.y;
+				if(diffX==0&&diffY==0)return;
+				final int passos=20;
+				final Point quadrante=new Point(localIni.x<=localFim.x?1:-1,localIni.y<localFim.y?1:-1);	//INF-DIR
+				final int fps=60;
+				final long tempoAlvo=1000000000/fps;
+				if(Math.abs(diffX)>=Math.abs(diffY)){
+					final double slope=diffY/diffX;
+					final double speed=Math.abs(diffX/passos);
+					for(double x=localIni.x;(quadrante.x==1?x<=localFim.x:x>localFim.x);x+=speed*quadrante.x){
+						final long tempoIni=System.nanoTime();
+						final double y=slope*(x-localIni.x)+localIni.y;	//y=m*x+b
+						Tree.setLocal((int)x,(int)y);
+						tree.draw();
+						if(animation!=this)return;	//OUTRA ANIMAÇÃO FOI CHAMADA
+						try{
+							Thread.sleep((tempoIni-System.nanoTime()+tempoAlvo)/1000000);
+						}catch(Exception erro){/*EVITA ERRO DE VALOR NEGATIVO*/}
+					}
+				}else{
+					final double slope=diffX/diffY;
+					final double speed=Math.abs(diffY/passos);
+					for(double y=localIni.y;(quadrante.y==1?y<=localFim.y:y>localFim.y);y+=speed*quadrante.y){
+						final long tempoIni=System.nanoTime();
+						final double x=slope*(y-localIni.y)+localIni.x;	//x=m*y+b
+						Tree.setLocal((int)x,(int)y);
+						tree.draw();
+						if(animation!=this)return;	//OUTRA ANIMAÇÃO FOI CHAMADA
+						try{
+							Thread.sleep((tempoIni-System.nanoTime()+tempoAlvo)/1000000);
+						}catch(Exception erro){/*EVITA ERRO DE VALOR NEGATIVO*/}
+					}
+				}
+				Tree.setLocal(localFim.x,localFim.y);
+				tree.draw();
+			}
+		}).start();
+	}
+	protected void startAutoDrag(){
 		new Thread(new Runnable(){
 			public void run(){
 				final int restRadius=40;
@@ -3159,7 +251,7 @@ public class Actions{
 				final long tempoAlvo=1000000000/fps;
 				double acumX=0;
 				double acumY=0;
-				while(getState()==AUTO_DRAG_ALL+NORMAL){
+				while(tree.getState()==TreeST.AUTO_DRAG_ALL+TreeST.NORMAL){
 					final long tempoIni=System.nanoTime();
 					if(!tree.isEnabled()){	//INTERROMPE SE DESFOCADO
 						try{
@@ -3201,7 +293,7 @@ public class Actions{
 		}).start();
 	}
 //GET OBJETO HOVER
-	private Objeto getObjetoHover(Point mouse){
+	protected Objeto getObjetoHover(Point mouse){
 		Objeto objHover=null;
 		if(objHover==null)for(Modulo mod:tree.getVisibleMods().values()){
 			if(mod.contains(mouse)&&mod!=Tree.getGhost())objHover=mod;
@@ -3214,7 +306,7 @@ public class Actions{
 		}
 		return objHover;
 	}
-	private Modulo getModuloHover(Point mouse,Modulo...modsToIgnore){
+	protected Modulo getModuloHover(Point mouse,Modulo...modsToIgnore){
 		Modulo modHover=null;
 		for(Modulo mod:tree.getVisibleMods().values()){
 			boolean toIgnore=false;
@@ -3229,14 +321,14 @@ public class Actions{
 		}
 		return modHover;
 	}
-	private Conexao getConexaoHover(Point mouse){
+	protected Conexao getConexaoHover(Point mouse){
 		Conexao coxHover=null;
 		for(Conexao cox:tree.getVisibleCoxs().values()){
 			if(cox.contains(mouse))coxHover=cox;
 		}
 		return coxHover;
 	}
-	private Nodulo getNoduloHover(Point mouse){
+	protected Nodulo getNoduloHover(Point mouse){
 		Nodulo nodHover=null;
 		for(Nodulo nod:tree.getVisibleNods().values()){
 			if(nod.contains(mouse))nodHover=nod;
@@ -3244,17 +336,18 @@ public class Actions{
 		return nodHover;
 	}
 //GET MOUSE POSITION
-	public Point getPosition(Point mouse){
+	protected Point getPosition(Point mouse){
+		final JFrame janela=(JFrame)tree.getPainel().getJanela();
 		final int menu=(janela.getJMenuBar()!=null?janela.getJMenuBar().getHeight():0);
 		return new Point(mouse.x-tree.getX()-janela.getInsets().left,mouse.y-tree.getY()-janela.getInsets().top-menu);
 	}
-	public Point getGridPosition(Point mouse){
+	protected Point getGridPosition(Point mouse){
 		mouse=getPosition(mouse);
 		mouse.x/=Tree.UNIT;
 		mouse.y/=Tree.UNIT;
 		return mouse;
 	}
-	public Point getNonGridPosition(Point mouse){
+	protected Point getNonGridPosition(Point mouse){
 		mouse=new Point(mouse);
 		mouse.x=(mouse.x*Tree.UNIT)+(Tree.UNIT/2);
 		mouse.y=(mouse.y*Tree.UNIT)+(Tree.UNIT/2);
@@ -3270,11 +363,11 @@ public class Actions{
 		return zoom;
 	}
 //TÍTULO
-	private void updateTituloBounds(){
-		if(tree.getTitulo().isVisible()){
-			final Modulo mod=(Modulo)tree.getTitulo().getObjeto();
+	protected void updateTituloBounds(){
+		if(tree.getUI().getTitulo().isVisible()){
+			final Modulo mod=(Modulo)tree.getUI().getTitulo().getObjeto();
 			final int iconSize=(mod.isIconified()?Icone.getSize(Tree.UNIT):0);
-			tree.getTitulo().setBounds(
+			tree.getUI().getTitulo().setBounds(
 					mod.getX(Tree.UNIT),
 					mod.getY(Tree.UNIT)+iconSize,
 					mod.getWidth(Tree.UNIT),
@@ -3284,30 +377,30 @@ public class Actions{
 	}
 //ATALHOS
 	public void setModo(int modo){
-		if(tree.getTitulo().isVisible())return;
-		final State state=getStates().get(getState());
+		if(tree.getUI().getTitulo().isVisible())return;
+		final TreeST.State state=tree.getStateContent(tree.getState());
 		switch(modo){
-			case TO_CREATE:
+			case TreeST.TO_CREATE:
 				if(state.getCreateAction()!=null)state.getCreateAction().run();
 			break;
-			case TO_CONNECT:
+			case TreeST.TO_CONNECT:
 				if(state.getConnectAction()!=null)state.getConnectAction().run();
 			break;
-			case TO_DELETE:
+			case TreeST.TO_DELETE:
 				if(state.getDeleteAction()!=null)state.getDeleteAction().run();
 			break;
 		}
 	}
 	//EXIBIR
 		public void centralizar(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			final Point oldLocal=Tree.getLocal();
 			if(tree.getSelectedObjetos().isEmpty()){
-				tree.setFocusOn(new Objeto[]{Tree.getMestre()});
+				setFocusOn(new Objeto[]{Tree.getMestre()});
 			}else{
-				tree.setFocusOn(tree.getSelectedObjetos().getAll().values().toArray(new Objeto[0]));
+				setFocusOn(tree.getSelectedObjetos().getAll().values().toArray(new Objeto[0]));
 			}
-			tree.animate(oldLocal,Tree.getLocal());
+			animate(oldLocal,Tree.getLocal());
 			resetState();
 		}
 		public void zoom(int rotation){
@@ -3316,23 +409,23 @@ public class Actions{
 		}
 	//EDITAR
 		public void undo(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			tree.getUndoRedoManager().undo();
 			resetState();
 		}
 		public void redo(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			tree.getUndoRedoManager().redo();
 			resetState();
 		}
 		public void cut(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			final String cut=String.join("\n",tree.getText(tree.getSelectedObjetos()));
 			Toolkit.getDefaultToolkit().getSystemClipboard().setContents(new StringSelection(cut),null);
 			delete();
 		}
 		public void copy(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			final String copy=String.join("\n",tree.getText(tree.getSelectedObjetos()));
 			Toolkit.getDefaultToolkit().getSystemClipboard().setContents(new StringSelection(copy),null);
 			resetState();
@@ -3358,14 +451,14 @@ public class Actions{
 			}
 		}
 		public void copyAsImg(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			final Image imagem=tree.getImage(new ListaObjeto(tree.getSelectedObjetos()));
 			if(imagem==null)return;
 			Toolkit.getDefaultToolkit().getSystemClipboard().setContents(new TransferableImage(imagem),null);
 			resetState();
 		}
 		public void paste(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			tree.unSelectAll();
 			try{
 				final String texto=Toolkit.getDefaultToolkit().getSystemClipboard().getData(DataFlavor.stringFlavor).toString();
@@ -3375,11 +468,11 @@ public class Actions{
 				final List<Objeto>lista=tree.addTree(mindTag,false);
 				for(Objeto obj:lista)tree.select(obj);
 			}catch(Exception erro){
-				Tree.mensagem(Tree.getLang().get("T_Err2","Error: Couldn't paste objects!")+"\n"+erro,Tree.Options.ERRO);
+				TreeUI.mensagem(TreeUI.getLang().get("T_Err2","Error: Couldn't paste objects!")+"\n"+erro,TreeUI.Options.ERRO);
 			}
 			if(tree.getSelectedObjetos().isEmpty())return;
 			final Point local=Tree.getLocal();
-			tree.setFocusOn(tree.getSelectedObjetos().getAll().values().toArray(new Objeto[0]));	//MOVE FOCO PARA O CENTRO DO GRUPO
+			setFocusOn(tree.getSelectedObjetos().getAll().values().toArray(new Objeto[0]));	//MOVE FOCO PARA O CENTRO DO GRUPO
 			final Point telaMeio=getGridPosition(new Point(tree.getPainel().getJanela().getWidth()/2,tree.getPainel().getJanela().getHeight()/2));
 				//A DIFERENÇA ENTRE OS LOCAIS ALINHA TUDO PARA O CENTRO DA JANELA, E A DIFERENÇA ENTRE O MOUSE E O CENTRO ALINHA TUDO PARA O CURSOR
 			final Point diff=new Point(local.x-Tree.getLocalX()-(mouseMoved.x-telaMeio.x),local.y-Tree.getLocalY()-(mouseMoved.y-telaMeio.y));
@@ -3391,9 +484,9 @@ public class Actions{
 			}
 			Tree.setLocal(local.x,local.y);
 			deleteGhostCoxs();
-			tree.getSelecao().setEmpty();
-			if(stateContains(TO_CREATE)){
-				final State state=getStates().get(getState());
+			tree.getUI().getSelecao().setEmpty();
+			if(tree.getST().stateContains(TreeST.TO_CREATE)){
+				final TreeST.State state=tree.getStateContent(tree.getState());
 				if(state.getModulo()!=null){
 					addUndoable(state.getModulo(),false);
 				}else if(state.getNodulo()!=null){
@@ -3401,30 +494,110 @@ public class Actions{
 				}
 			}
 			tree.draw();
-			setState(MOVE+NORMAL);
+			tree.setState(TreeST.MOVE+TreeST.NORMAL);
 		}
 		public void delete(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			tree.unSelect(Tree.getMestre());
 			if(tree.getSelectedObjetos().isEmpty())return;
 			delUndoable(new ArrayList<>(tree.getSelectedObjetos().getAll().values()));
 			resetState();
 		}
+		public void moveUp(boolean saltar){
+			if(tree.getUI().getTitulo().isVisible())return;
+			if(tree.getUI().getPopup().isShowing())return;
+			final int salto=(saltar?8:1);
+			if(tree.getSelectedObjetos().isEmpty()){
+				Tree.setLocal(Tree.getLocalX(),Tree.getLocalY()+salto);
+			}else{
+				for(Modulo mod:tree.getSelectedObjetos().getModulos()){
+					mod.setLocationIndex(mod.getXIndex(),mod.getYIndex()-salto);
+				}
+				for(Nodulo nod:tree.getSelectedObjetos().getNodulos()){
+					nod.setLocationIndex(nod.getXIndex(),nod.getYIndex()-salto);
+				}
+				final List<Objeto>savedObjs=new ArrayList<>();
+				savedObjs.addAll(tree.getSelectedObjetos().getModulos());
+				savedObjs.addAll(tree.getSelectedObjetos().getNodulos());
+				tree.getUndoRedoManager().addUndoLocal(savedObjs.toArray(new Objeto[0]),new Point(0,+salto));
+			}
+			tree.draw();
+		}
+		public void moveRight(boolean saltar){
+			if(tree.getUI().getTitulo().isVisible())return;
+			if(tree.getUI().getPopup().isShowing())return;
+			final int salto=(saltar?8:1);
+			if(tree.getSelectedObjetos().isEmpty()){
+				Tree.setLocal(Tree.getLocalX()-salto,Tree.getLocalY());
+			}else{
+				for(Modulo mod:tree.getSelectedObjetos().getModulos()){
+					mod.setLocationIndex(mod.getXIndex()+salto,mod.getYIndex());
+				}
+				for(Nodulo nod:tree.getSelectedObjetos().getNodulos()){
+					nod.setLocationIndex(nod.getXIndex()+salto,nod.getYIndex());
+				}
+				final List<Objeto>savedObjs=new ArrayList<>();
+				savedObjs.addAll(tree.getSelectedObjetos().getModulos());
+				savedObjs.addAll(tree.getSelectedObjetos().getNodulos());
+				tree.getUndoRedoManager().addUndoLocal(savedObjs.toArray(new Objeto[0]),new Point(-salto,0));
+			}
+			tree.draw();
+		}
+		public void moveLeft(boolean saltar){
+			if(tree.getUI().getTitulo().isVisible())return;
+			if(tree.getUI().getPopup().isShowing())return;
+			final int salto=(saltar?8:1);
+			if(tree.getSelectedObjetos().isEmpty()){
+				Tree.setLocal(Tree.getLocalX()+salto,Tree.getLocalY());
+			}else{
+				for(Modulo mod:tree.getSelectedObjetos().getModulos()){
+					mod.setLocationIndex(mod.getXIndex()-salto,mod.getYIndex());
+				}
+				for(Nodulo nod:tree.getSelectedObjetos().getNodulos()){
+					nod.setLocationIndex(nod.getXIndex()-salto,nod.getYIndex());
+				}
+				final List<Objeto>savedObjs=new ArrayList<>();
+				savedObjs.addAll(tree.getSelectedObjetos().getModulos());
+				savedObjs.addAll(tree.getSelectedObjetos().getNodulos());
+				tree.getUndoRedoManager().addUndoLocal(savedObjs.toArray(new Objeto[0]),new Point(+salto,0));
+			}
+			tree.draw();
+		}
+		public void moveDown(boolean saltar){
+			if(tree.getUI().getTitulo().isVisible())return;
+			if(tree.getUI().getPopup().isShowing())return;
+			final int salto=(saltar?8:1);
+			if(tree.getSelectedObjetos().isEmpty()){
+				Tree.setLocal(Tree.getLocalX(),Tree.getLocalY()-salto);
+			}else{
+				for(Modulo mod:tree.getSelectedObjetos().getModulos()){
+					mod.setLocationIndex(mod.getXIndex(),mod.getYIndex()+salto);
+				}
+				for(Nodulo nod:tree.getSelectedObjetos().getNodulos()){
+					nod.setLocationIndex(nod.getXIndex(),nod.getYIndex()+salto);
+				}
+				final List<Objeto>savedObjs=new ArrayList<>();
+				savedObjs.addAll(tree.getSelectedObjetos().getModulos());
+				savedObjs.addAll(tree.getSelectedObjetos().getNodulos());
+				tree.getUndoRedoManager().addUndoLocal(savedObjs.toArray(new Objeto[0]),new Point(0,-salto));
+			}
+			tree.draw();
+		}
 	//SELECIONAR
 		public void selectAll(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			tree.selectAll();
 			resetState();
 		}
 		public void unSelectAll(){
-			if(tree.getTitulo().isVisible()){
-				tree.setTitulo(null);
+			if(tree.getUI().getTitulo().isVisible()){
+				tree.getUI().setTitulo(null);
 			}
 			tree.unSelectAll();
 			resetState();
 		}
 		public void selectSons(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			final List<Objeto>treesSelec=new ArrayList<>();
 			treesSelec.addAll(tree.getSelectedObjetos().getModulos());
 			treesSelec.addAll(tree.getSelectedObjetos().getConexoes());
@@ -3432,14 +605,14 @@ public class Actions{
 			resetState();
 		}
 		public void invertSelection(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			final List<Objeto>objs=new ArrayList<>(tree.getSelectedObjetos().getAll().values());
 			tree.selectAll();
 			for(Objeto obj:objs)tree.unSelect(obj);
 			resetState();
 		}
 		public void selectModSemPai(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			tree.unSelectAll();
 			for(Objeto obj:tree.getObjetos().getAll().values()){
 				if(obj.getTipo().is(Objeto.Tipo.MODULO)){
@@ -3459,42 +632,42 @@ public class Actions{
 		}
 	//MÓDULO
 		public void editTitulo(){
-			if(getState()!=NORMAL)return;
+			if(tree.getState()!=TreeST.NORMAL)return;
 			if(tree.getSelectedObjetos().getModulos().size()!=1)return;
-			tree.setTitulo(tree.getSelectedObjetos().getModulos().get(0));
-			setState(Actions.EDIT_TITLE);
+			tree.getUI().setTitulo(tree.getSelectedObjetos().getModulos().get(0));
+			tree.setState(TreeST.EDIT_TITLE);
 		}
 		public void createModRelacionado(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			deleteGhostCoxs();
-			tree.getSelecao().setEmpty();
+			tree.getUI().getSelecao().setEmpty();
 			if(!tree.getSelectedObjetos().getModulos().isEmpty()){
 				final Modulo newMod=createModRelacionado(
 						mouseMoved.x-Tree.getLocalX(),mouseMoved.y-Tree.getLocalY(),
 						tree.getSelectedObjetos().getModulos());
-				setState(MOVE+TO_CREATE,newMod);
+				tree.setState(TreeST.MOVE+TreeST.TO_CREATE,newMod);
 			}else{
 				tree.unSelectAll();
 				tree.draw();
-				setState(NORMAL);
+				tree.setState(TreeST.NORMAL);
 			}
 		}
 		public void startRelation(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			deleteGhostCoxs();
-			tree.getSelecao().setEmpty();
+			tree.getUI().getSelecao().setEmpty();
 			Tree.getGhost().setLocationIndex(mouseMoved.x-Tree.getLocalX(),mouseMoved.y-Tree.getLocalY());
 			relateToGhost(tree.getSelectedObjetos().getModulos());
 			tree.unSelectAll();
 			tree.draw();
 			if(!Tree.getGhost().getConexoes().isEmpty()){
-				setState(MOVE+TO_CONNECT);
-			}else setState(NORMAL);
+				tree.setState(TreeST.MOVE+TreeST.TO_CONNECT);
+			}else tree.setState(TreeST.NORMAL);
 		}
 		public void deleteMods(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			deleteGhostCoxs();
-			tree.getSelecao().setEmpty();
+			tree.getUI().getSelecao().setEmpty();
 			tree.unSelect(Tree.getMestre());
 			if(tree.getSelectedObjetos().getModulos().isEmpty())return;
 			final List<Modulo>selectedMods=new ArrayList<>(tree.getSelectedObjetos().getModulos());
@@ -3509,7 +682,7 @@ public class Actions{
 		}
 	//CONEXÃO
 		public void invertCox(){
-			if(tree.getTitulo().isVisible())return;
+			if(tree.getUI().getTitulo().isVisible())return;
 			final List<Objeto>newObjs=new ArrayList<>();
 			for(Conexao cox:tree.getSelectedObjetos().getConexoes()){
 				if(cox.getPai()==Tree.getMestre())continue;
@@ -3533,9 +706,9 @@ public class Actions{
 //RESET
 	public void resetState(){
 		deleteGhostCoxs();
-		tree.getSelecao().setEmpty();
-		if(stateContains(TO_CREATE)){
-			final State state=getStates().get(getState());
+		tree.getUI().getSelecao().setEmpty();
+		if(tree.getST().stateContains(TreeST.TO_CREATE)){
+			final TreeST.State state=tree.getStateContent(tree.getState());
 			if(state.getModulo()!=null){
 				addUndoable(state.getModulo(),false);
 			}else if(state.getNodulo()!=null){
@@ -3543,7 +716,7 @@ public class Actions{
 			}
 		}
 		tree.draw();
-		setState(NORMAL);
+		tree.setState(TreeST.NORMAL);
 	}
 //REPETIÇÕES
 	//ADD/DEL
@@ -3595,10 +768,10 @@ public class Actions{
 			delUndoable(objs);
 		}
 	//SET
-		private void setHover(Selecao.State state){
+		protected void setHover(Selecao.State state){
 			final Point mouse=getNonGridPosition(mouseMoved);
 			tree.unSelectAll();
-			tree.getPopup().close();
+			tree.getUI().getPopup().close();
 			Modulo modHover=getModuloHover(mouse);
 			switch(state){
 				case TO_UNSELECT:default:break;
@@ -3642,7 +815,7 @@ public class Actions{
 			}
 			tree.draw();
 		}
-		private void setArea(Selecao.State state,boolean addTo,int...locations){
+		protected void setArea(Selecao.State state,boolean addTo,int...locations){
 			Rectangle allChunksIndexes=null;
 			Rectangle newChunksIndexes=null;
 			switch(locations.length){
@@ -3652,8 +825,8 @@ public class Actions{
 						case TO_DELETE:	tree.unSelectAll();			break;
 						default:		tree.unSelectAllMods();		break;
 					}
-					tree.getSelecao().setState(state);
-					allChunksIndexes=tree.getChunksIndexes(tree.getSelecao().getFormIndex());
+					tree.getUI().getSelecao().setState(state);
+					allChunksIndexes=tree.getChunksIndexes(tree.getUI().getSelecao().getFormIndex());
 					newChunksIndexes=allChunksIndexes;
 				break;
 				case 4:		//CRIA UMA ÁREA
@@ -3662,16 +835,16 @@ public class Actions{
 						case TO_DELETE:	tree.unSelectAll();			break;
 						default:		tree.unSelectAllMods();		break;
 					}
-					tree.getSelecao().setState(state);
-					tree.getSelecao().setAncoraIndex(locations[0],locations[1]);
-					tree.getSelecao().setAreaIndex(locations[2],locations[3]);
-					allChunksIndexes=tree.getChunksIndexes(tree.getSelecao().getFormIndex());
+					tree.getUI().getSelecao().setState(state);
+					tree.getUI().getSelecao().setAncoraIndex(locations[0],locations[1]);
+					tree.getUI().getSelecao().setAreaIndex(locations[2],locations[3]);
+					allChunksIndexes=tree.getChunksIndexes(tree.getUI().getSelecao().getFormIndex());
 					newChunksIndexes=allChunksIndexes;
 				break;
 				case 2:		//EXPANDE UMA ÁREA
-					final Rectangle areaOld=tree.getSelecao().getFormIndex();
-					tree.getSelecao().setAreaIndex(locations[0],locations[1]);
-					final Rectangle areaNew=tree.getSelecao().getFormIndex();
+					final Rectangle areaOld=tree.getUI().getSelecao().getFormIndex();
+					tree.getUI().getSelecao().setAreaIndex(locations[0],locations[1]);
+					final Rectangle areaNew=tree.getUI().getSelecao().getFormIndex();
 					final Rectangle areaAll=new Rectangle(		//ÁREA TOTAL
 							Math.min(areaOld.x,areaNew.x),
 							Math.min(areaOld.y,areaNew.y),
@@ -3698,7 +871,7 @@ public class Actions{
 							}
 						}else{
 							for(Modulo mod:chunk.getObjetos().getModulos()){	//SELEC MODS
-								if(tree.getSelecao().intersects(mod.getForm(Tree.UNIT))){
+								if(tree.getUI().getSelecao().intersects(mod.getForm(Tree.UNIT))){
 									switch(state){
 										case TO_SELECT:default:	tree.select(mod);				break;
 										case TO_CREATE:			tree.selectToBeCreator(mod);	break;
@@ -3709,7 +882,7 @@ public class Actions{
 								}else if(!addTo)tree.unSelect(mod);
 							}
 							for(Conexao cox:chunk.getObjetos().getConexoes()){	//SELEC COXS
-								if(tree.getSelecao().intersects(cox.getForm(Tree.UNIT))){
+								if(tree.getUI().getSelecao().intersects(cox.getForm(Tree.UNIT))){
 									switch(state){
 										case TO_SELECT:default:	tree.select(cox);				break;
 										case TO_CREATE:			break;
@@ -3724,7 +897,7 @@ public class Actions{
 								}
 							}
 							for(Nodulo nod:chunk.getObjetos().getNodulos()){	//SELEC NODS
-								if(tree.getSelecao().intersects(nod.getForm(Tree.UNIT))){
+								if(tree.getUI().getSelecao().intersects(nod.getForm(Tree.UNIT))){
 									switch(state){
 										case TO_SELECT:default:	tree.select(nod);				break;
 										case TO_CREATE:			break;
@@ -3744,7 +917,7 @@ public class Actions{
 			}
 			tree.draw();
 		}
-		private void addAndSelect(Objeto obj){
+		protected void addAndSelect(Objeto obj){
 			if(obj==null)return;
 			switch(obj.getTipo()){
 				case MODULO:
@@ -3801,7 +974,7 @@ public class Actions{
 			tree.draw();
 			return newNod;
 		}
-		private void createNodOnGhostCoxs(int x,int y){
+		protected void createNodOnGhostCoxs(int x,int y){
 			for(Conexao cox:Tree.getGhost().getConexoes()){
 				final Nodulo newNod=new Nodulo(cox,x,y);
 				tree.add(newNod);
@@ -3809,7 +982,7 @@ public class Actions{
 			tree.draw();
 		}
 	//RELATE
-		private void relateMods(List<Modulo>modsPais,Modulo modSon){
+		protected void relateMods(List<Modulo>modsPais,Modulo modSon){
 			for(Modulo modPai:modsPais){
 				if(modPai==modSon)continue;
 				final Conexao newCox=new Conexao(modPai,modSon);
@@ -3820,7 +993,7 @@ public class Actions{
 			tree.unSelectAll();
 			tree.draw();
 		}
-		private void relateToMods(List<Modulo>mods){
+		protected void relateToMods(List<Modulo>mods){
 			if(Tree.getGhost().getConexoes().isEmpty())return;
 			final List<Objeto>savedObjs=new ArrayList<>();
 			for(Modulo mod:mods){
@@ -3841,7 +1014,7 @@ public class Actions{
 			for(Objeto obj:savedObjs)tree.select(obj);
 			tree.draw();
 		}
-		private void relateToMods(Modulo mod){relateToMods(Arrays.asList(mod));}
+		protected void relateToMods(Modulo mod){relateToMods(Arrays.asList(mod));}
 		public void relateToGhost(List<Modulo>mods){
 			for(Modulo mod:mods){
 				final Conexao newCox=new Conexao(Tree.getGhost(),mod);
@@ -3851,9 +1024,9 @@ public class Actions{
 			tree.unSelectAll();
 			tree.draw();
 		}
-		private void relateToGhost(Modulo mod){relateToGhost(Arrays.asList(mod));}
+		public void relateToGhost(Modulo mod){relateToGhost(Arrays.asList(mod));}
 	//DELETE
-		private void deleteGhostCoxs(){
+		public void deleteGhostCoxs(){
 			while(!Tree.getGhost().getConexoes().isEmpty()){
 				tree.del(Tree.getGhost().getConexoes().get(0));
 			}
