+++ utilitarios/yo/element/tree/objeto/conexao/Conexao.java
@@ -1,119 +1,34 @@
 package element.tree.objeto.conexao;
-import java.awt.BasicStroke;
 import java.awt.Graphics2D;
 import java.awt.Point;
 import java.awt.Rectangle;
-import java.awt.geom.Ellipse2D;
 import java.awt.geom.Line2D;
 import java.awt.geom.Path2D;
-import java.awt.geom.Point2D;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
-
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
-
 import element.tree.propriedades.Borda;
-import element.tree.propriedades.Cor;
 import element.tree.propriedades.Grossura;
 import element.tree.Tree;
 import element.tree.objeto.Objeto;
 import element.tree.objeto.ObjetoBoundsListener;
+import element.tree.objeto.nodulo.Nodulo;
+import element.tree.objeto.conexao.segmento.Segmento;
 import element.tree.objeto.modulo.Modulo;
 public class Conexao extends Objeto{
-//STATES
-	public enum State{
-		UNSELECTED,
-		SELECTED,
-		HIGHLIGHTED,
-		TO_BE_CREATOR,
-		TO_BE_DELETED;
-		public boolean is(Conexao.State... states){
-			for(Conexao.State state:states)if(this.equals(state))return true;
-			return false;
-		}
-	}
-	private Conexao.State state=Conexao.State.UNSELECTED;
-		public Conexao.State getState(){return state;}
-		public void setState(Conexao.State state){
-			switch(getState()){
-				case UNSELECTED:
-					if(state.is(Conexao.State.SELECTED)){
-						this.state=state;							//UNSELECTED -> SELECTED
-						for(Nodulo nod:getNodulos())nod.setState(Nodulo.State.HIGHLIGHTED);
-						getPai().setState(Modulo.State.HIGHLIGHTED);
-						getSon().setState(Modulo.State.HIGHLIGHTED);
-					}else if(state.is(Conexao.State.HIGHLIGHTED)){
-						this.state=state;							//UNSELECTED -> HIGHLIGHTED
-						for(Nodulo nod:getNodulos())nod.setState(Nodulo.State.HIGHLIGHTED);
-						getPai().setState(Modulo.State.HIGHLIGHTED);
-						getSon().setState(Modulo.State.HIGHLIGHTED);
-					}else if(state.is(Conexao.State.TO_BE_CREATOR)){
-						this.state=state;							//UNSELECTED -> TO_BE_CREATOR
-						for(Nodulo nod:getNodulos())nod.setState(Nodulo.State.TO_BE_CREATOR);
-					}else if(state.is(Conexao.State.TO_BE_DELETED)){
-						this.state=state;							//UNSELECTED -> TO_BE_DELETED
-						for(Nodulo nod:getNodulos())nod.setState(Nodulo.State.TO_BE_DELETED);
-					}
-				break;
-				case SELECTED:
-					if(state.is(Conexao.State.UNSELECTED)){
-						if(getPai().getState().is(Modulo.State.SELECTED)||getSon().getState().is(Modulo.State.SELECTED)){
-							this.state=Conexao.State.HIGHLIGHTED;	//SELECTED -> HIGHLIGHTED
-							for(Nodulo nod:getNodulos())nod.setState(Nodulo.State.HIGHLIGHTED);
-							getPai().setState(Modulo.State.HIGHLIGHTED);
-							getSon().setState(Modulo.State.HIGHLIGHTED);
-						}else{
-							this.state=state;						//SELECTED -> UNSELECTED
-							for(Nodulo nod:getNodulos())if(nod.getState().is(Nodulo.State.HIGHLIGHTED))nod.setState(Nodulo.State.UNSELECTED);
-							if(getPai().getState().is(Modulo.State.HIGHLIGHTED))getPai().setState(Modulo.State.UNSELECTED);
-							if(getSon().getState().is(Modulo.State.HIGHLIGHTED))getSon().setState(Modulo.State.UNSELECTED);
-						}
-					}
-				break;
-				case HIGHLIGHTED:
-					if(state.is(Conexao.State.SELECTED)){
-						this.state=state;							//HIGHLIGHTED -> SELECTED
-						for(Nodulo nod:getNodulos())nod.setState(Nodulo.State.HIGHLIGHTED);
-						getPai().setState(Modulo.State.HIGHLIGHTED);
-						getSon().setState(Modulo.State.HIGHLIGHTED);
-					}else if(state.is(Conexao.State.UNSELECTED)){
-						this.state=state;							//HIGHLIGHTED -> UNSELECTED
-						if(getPai().getState().is(Modulo.State.SELECTED)||getSon().getState().is(Modulo.State.SELECTED)){
-							this.state=Conexao.State.HIGHLIGHTED;	//UNSELECTED -> HIGHLIGHTED
-						}
-						if(getState().is(Conexao.State.UNSELECTED))for(Nodulo nod:getNodulos())if(nod.getState().is(Nodulo.State.SELECTED)){
-							this.state=Conexao.State.HIGHLIGHTED;	//UNSELECTED -> HIGHLIGHTED
-							break;
-						}
-						if(getState().is(Conexao.State.UNSELECTED)){
-							for(Nodulo nod:getNodulos())nod.setState(Nodulo.State.UNSELECTED);
-							getPai().setState(Modulo.State.UNSELECTED);
-							getSon().setState(Modulo.State.UNSELECTED);
-						}else{
-							for(Nodulo nod:getNodulos())nod.setState(Nodulo.State.HIGHLIGHTED);
-							getPai().setState(Modulo.State.HIGHLIGHTED);
-							getSon().setState(Modulo.State.HIGHLIGHTED);
-						}
-					}
-				break;
-				case TO_BE_CREATOR:
-					if(state.is(Conexao.State.UNSELECTED)){
-						this.state=state;							//TO_BE_CREATOR -> UNSELECTED
-						for(Nodulo nod:getNodulos())nod.setState(Nodulo.State.UNSELECTED);
-						for(Segmento seg:getSegmentos())seg.setState(Segmento.State.UNSELECTED);
-					}
-				break;
-				case TO_BE_DELETED:
-					if(state.is(Conexao.State.UNSELECTED)){
-						if(getPai().getState().is(Modulo.State.TO_BE_DELETED)||getSon().getState().is(Modulo.State.TO_BE_DELETED))return;
-						this.state=state;							//TO_BE_DELETED -> UNSELECTED
-						for(Nodulo nod:getNodulos())nod.setState(Nodulo.State.UNSELECTED);
-					}
-				break;
-			}
+//ST
+	private final ConexaoST ST=new ConexaoST(this);
+		public ConexaoST getST(){return ST;}
+		public ConexaoST.State getState(){return getST().getState();}
+		public void setState(ConexaoST.State state){getST().setState(state);}
+//UI
+	private final ConexaoUI UI=new ConexaoUI(this);
+		public ConexaoUI getUI(){return UI;}
+		public void draw(Graphics2D imagemEdit,int unit){
+			getUI().draw(imagemEdit,unit);
 		}
 //LOCAL
 	public int getX1Index(){return son.getMeioXIndex();}
@@ -305,8 +220,9 @@ public class Conexao extends Objeto{
 	public int hashCode(){	//ALGUMAS PARTES SÃO DESNECESSÁRIAS PARA DISTINÇÃO
 		final int prime=31;
 		int result=super.hashCode();
-		result=prime*result+getIndex();
-//		result=prime*result+((state==null)?0:state.hashCode());
+		//TODO: ADICIONAR INDEX
+//		result=prime*result+getIndex();
+//		result=prime*result+((ST==null)?0:ST.hashCode());
 		result=prime*result+((pai==null)?0:pai.hashCode());
 		result=prime*result+((son==null)?0:son.hashCode());
 //		result=prime*result+((nodulos==null)?0:nodulos.hashCode());
@@ -323,7 +239,7 @@ public class Conexao extends Objeto{
 		if(!super.equals(obj))return false;
 		if(getClass()!=obj.getClass())return false;
 		final Conexao cox=(Conexao)obj;
-//		if(state!=cox.state)return false;
+//		if(ST!=cox.ST)return false;
 		if(pai!=cox.pai)return false;
 		if(son!=cox.son)return false;
 //		if(nodulos!=cox.nodulos)return false;
@@ -332,155 +248,6 @@ public class Conexao extends Objeto{
 //		if(texto!=cox.texto)return false;
 //		if(caret!=cox.caret)return false;
 		return true;
-	}
-//DRAW
-	public void draw(Graphics2D imagemEdit,int unit){
-		final float borda=Tree.getBordaValue(unit);
-		drawBrilho(imagemEdit,unit,borda);
-		drawLinha(imagemEdit,unit,borda);
-		drawPonta(imagemEdit,unit);
-	}
-		private void drawBrilho(Graphics2D imagemEdit,int unit,float borda){
-			if(unit<=2)return;		//EM ZOOM<=2, É IRRELEVANTE
-			if(getState().is(Conexao.State.UNSELECTED,Conexao.State.HIGHLIGHTED))return;
-			switch(getGrossura().getIndex()){
-				case Grossura.THIN:
-					imagemEdit.setStroke(new BasicStroke(getGrossura().getValor(borda)*4*4,BasicStroke.CAP_ROUND,BasicStroke.JOIN_BEVEL));
-				break;
-				case Grossura.MEDIUM:default:
-					imagemEdit.setStroke(new BasicStroke(getGrossura().getValor(borda)*3*3,BasicStroke.CAP_ROUND,BasicStroke.JOIN_BEVEL));
-				break;
-				case Grossura.WIDE:
-					imagemEdit.setStroke(new BasicStroke(getGrossura().getValor(borda)*2*2,BasicStroke.CAP_ROUND,BasicStroke.JOIN_BEVEL));
-				break;
-				case Grossura.ULTRA_WIDE:
-					imagemEdit.setStroke(new BasicStroke(getGrossura().getValor(borda)*1.5f*1.5f,BasicStroke.CAP_ROUND,BasicStroke.JOIN_BEVEL));
-				break;
-			}
-			imagemEdit.setColor(getBrilhoCor(getState()));
-			if(getState().is(Conexao.State.TO_BE_CREATOR)){		//BRILHO DO SEG TO_BE_CREATOR
-				for(Segmento seg:getSegmentos()){
-					if(seg.getState().is(Segmento.State.UNSELECTED))continue;
-					final Path2D.Float forma=new Path2D.Float();
-					forma.moveTo(seg.getX1(unit),seg.getY1(unit));
-					forma.lineTo(seg.getX2(unit),seg.getY2(unit));
-					imagemEdit.draw(forma);
-				}
-				return;
-			}
-			final Path2D.Float forma=new Path2D.Float();
-			forma.moveTo(getX1(unit),getY1(unit));
-			for(Segmento seg:getSegmentos()){
-				if(seg.getPonta2()==Tree.getGhost()){
-					forma.lineTo(Tree.getGhost().getX(unit),Tree.getGhost().getY(unit));
-				}else forma.lineTo(seg.getX2(unit),seg.getY2(unit));
-			}
-			imagemEdit.draw(forma);								//BRILHO DO SELECTED OU TO_BE_DELETED
-		}
-		private void drawLinha(Graphics2D imagemEdit,int unit,float borda){
-			if(unit<=2&&getGrossura().getIndex()==Grossura.THIN)return;		//EM ZOOM<=2 COM GROSSURA FINA, É IRRELEVANTE
-			if(unit<=1&&getGrossura().getIndex()==Grossura.MEDIUM)return;	//EM ZOOM<=1 COM GROSSURA MEDIUM, É IRRELEVANTE
-			imagemEdit.setStroke(getBorda().getVisual(getGrossura().getValor(borda)));
-			imagemEdit.setColor(getLinhaCor(getState()));
-			final Path2D.Float forma=new Path2D.Float();
-			forma.moveTo(getX1(unit),getY1(unit));
-			for(Segmento seg:getSegmentos()){
-				if(seg.getPonta2()==Tree.getGhost()){
-					forma.lineTo(Tree.getGhost().getX(unit),Tree.getGhost().getY(unit));
-				}else forma.lineTo(seg.getX2(unit),seg.getY2(unit));
-			}
-			imagemEdit.draw(forma);
-		}
-		private void drawPonta(Graphics2D imagemEdit,int unit){
-			if(unit<=2)return;		//EM ZOOM<=2, É IRRELEVANTE
-			float x1=getX1(unit);
-			float y1=getY1(unit);
-			float x2=getX2(unit);
-			float y2=getY2(unit);
-			if(getPai()==Tree.getGhost()){
-				x2=Tree.getGhost().getX(unit);
-				y2=Tree.getGhost().getY(unit);
-			}
-			if(!getNodulos().isEmpty()){
-				final Nodulo nod=getNodulos().get(getNodulos().size()-1);
-				x1=nod.getX(unit);
-				y1=nod.getY(unit);
-			}
-			final Point2D.Float ponta=getPonta(unit,x1,y1,x2,y2);
-			float size=unit;
-			switch(getGrossura().getIndex()){
-				case Grossura.THIN:				size*=0.5f;	break;
-				case Grossura.MEDIUM:default:	size*=1.0f;	break;
-				case Grossura.WIDE:				size*=2.0f;	break;
-				case Grossura.ULTRA_WIDE:		size*=3.0f;	break;
-			}
-			final float x=(float)(ponta.getX())-(size/2);
-			final float y=(float)(ponta.getY())-(size/2);
-			imagemEdit.fill(new Ellipse2D.Float(x,y,size,size));
-		}
-			private Point2D.Float getPonta(int unit,float x1,float y1,float x2,float y2){
-				float x=0;
-				float y=0;
-				if(getPai()==Tree.getGhost()){
-					x=x2;
-					y=y2;
-				}else{
-					if(y1>y2){
-						if(x1>x2){	//DIR-INF
-							x=getPai().getX(unit)+getPai().getWidth(unit);
-							y=y2+(((y1-y2)*(getPai().getX(unit)+getPai().getWidth(unit)-x2))/Math.max(1,x1-x2));
-							if(y>getPai().getY(unit)+getPai().getHeight(unit)){	//INF-DIR
-								x=x2+(((x1-x2)*(getPai().getY(unit)+getPai().getHeight(unit)-y2))/Math.max(1,y1-y2));
-								y=getPai().getY(unit)+getPai().getHeight(unit);
-							}
-						}else{		//INF-ESQ
-							x=x2-(((x2-x1)*(getPai().getY(unit)+getPai().getHeight(unit)-y2))/Math.max(1,y1-y2));
-							y=getPai().getY(unit)+getPai().getHeight(unit);
-							if(x<getPai().getX(unit)){	//ESQ-INF
-								x=getPai().getX(unit);
-								y=y2+(((y1-y2)*(x2-getPai().getX(unit)))/Math.max(1,x2-x1));
-							}
-						}
-					}else{
-						if(x1<x2){	//ESQ-SUP
-							x=getPai().getX(unit);
-							y=y2-(((y2-y1)*(x2-getPai().getX(unit)))/Math.max(1,x2-x1));
-							if(y<getPai().getY(unit)){	//SUP-ESQ
-								x=x2-(((x2-x1)*(y2-getPai().getY(unit)))/Math.max(1,y2-y1));
-								y=getPai().getY(unit);
-							}
-						}else{
-							if(y2==y1)x=getPai().getX(unit)+getPai().getWidth(unit)+1;else{	//SUP-DIR
-								x=x2+(((x1-x2)*(y2-getPai().getY(unit)))/Math.max(1,y2-y1));
-								y=getPai().getY(unit);
-							}
-							if(x>getPai().getX(unit)+getPai().getWidth(unit)){	//DIR-SUP
-								x=getPai().getX(unit)+getPai().getWidth(unit);
-								y=y2-(((y2-y1)*(getPai().getX(unit)+getPai().getWidth(unit)-x2))/Math.max(1,x1-x2));
-							}
-						}
-					}
-				}
-				return new Point2D.Float(x,y);
-			}
-	public Cor getBrilhoCor(Conexao.State state){
-		switch(state){
-			case SELECTED:default:	return getSon().getBrilhoCor(Modulo.State.SELECTED);
-			case TO_BE_CREATOR:		return getSon().getBrilhoCor(Modulo.State.TO_BE_CREATOR);
-			case TO_BE_DELETED:		return getSon().getBrilhoCor(Modulo.State.TO_BE_DELETED);
-		}
-	}
-	public Cor getLinhaCor(Conexao.State state){
-		switch(state){
-			case UNSELECTED:default:	return getSon().getBordaCor(Modulo.State.UNSELECTED);
-			case SELECTED:				return getSon().getBordaCor(Modulo.State.SELECTED);
-			case HIGHLIGHTED:
-				if(getSon().getState().is(Modulo.State.SELECTED)){
-					return getSon().getBordaCor(Modulo.State.SELECTED);
-				}else return getSon().getBordaCor(Modulo.State.HIGHLIGHTED);
-			case TO_BE_CREATOR:			return getSon().getBordaCor(Modulo.State.TO_BE_CREATOR);
-			case TO_BE_DELETED:			return getSon().getBordaCor(Modulo.State.TO_BE_DELETED);
-		}
 	}
 //TAG -> COX
 	public static Conexao getConexao(Element coxTag,HashMap<Integer,Modulo>ids){
