+++ utilitarios/yo/element/tree/popup/icon/IconeBloco.java
@@ -8,6 +8,7 @@ import element.tree.propriedades.Cor;
 import element.tree.Cursor;
 import element.tree.propriedades.Icone;
 import element.tree.objeto.modulo.Modulo;
+import element.tree.objeto.modulo.ModuloUI;
 import element.tree.popup.Bloco;
 import element.tree.popup.Popup;
 @SuppressWarnings("serial")
@@ -77,7 +78,7 @@ public class IconeBloco extends Bloco{
 //DRAW
 	public void draw(Graphics imagemEdit){
 		if(isFocused()){
-			imagemEdit.setColor(Cor.getChanged(Modulo.Cores.FUNDO,1.5f));
+			imagemEdit.setColor(Cor.getChanged(ModuloUI.Cores.FUNDO,1.5f));
 			imagemEdit.fillRect((int)getX(),(int)getY(),(int)getWidth(),(int)getHeight());
 		}
 		imagemEdit.drawImage(	//INSERE ICONE CENTRALIZADO DO BLOCO
